arrow::array::array

Struct NativeAdapter

pub struct NativeAdapter<T>
where
    T: ArrowPrimitiveType,
{
    pub native: Option<<T as ArrowPrimitiveType>::Native>,
}

An optional primitive value

This struct is used as an adapter when creating PrimitiveArray from an iterator. FromIterator for PrimitiveArray takes an iterator where the elements can be into this struct. So once implementing From or Into trait for a type, an iterator of the type can be collected to PrimitiveArray.

Fields
native: Option<<T as ArrowPrimitiveType>::Native>

Corresponding Rust native type if available