arrow::array::array

Struct StructArray

pub struct StructArray {
    len: usize,
    data_type: DataType,
    nulls: Option<NullBuffer>,
    fields: Vec<Arc<dyn Array>>,
}

An array of structs

Each child (called field) is represented by a separate array.

Comparison with RecordBatch

Both RecordBatch and StructArray represent a collection of columns / arrays with the same length.

However, there are a couple of key differences:

StructArray can be nested within other Array, including itself

RecordBatch can contain top-level metadata on its associated Schema

StructArray can contain top-level nulls, i.e. null

RecordBatch can only represent nulls in its child columns, i.e. {"field": null}

StructArray is therefore a more general data container than RecordBatch, and as such code that needs to handle both will typically share an implementation in terms of StructArray and convert to/from RecordBatch as necessary.

From implementations are provided to facilitate this conversion, however, converting from a StructArray containing top-level nulls to a RecordBatch will panic, as there is no way to preserve them.

Example: Create an array from a vector of fields

use std::sync::Arc;
use arrow_array::{Array, ArrayRef, BooleanArray, Int32Array, StructArray};
use arrow_schema::{DataType, Field};

let boolean = Arc::new(BooleanArray::from(vec![false, false, true, true]));
let int = Arc::new(Int32Array::from(vec![42, 28, 19, 31]));

let struct_array = StructArray::from(vec![
    (
        Arc::new(Field::new("b", DataType::Boolean, false)),
        boolean.clone() as ArrayRef,
    ),
    (
        Arc::new(Field::new("c", DataType::Int32, false)),
        int.clone() as ArrayRef,
    ),
]);
assert_eq!(struct_array.column(0).as_ref(), boolean.as_ref());
assert_eq!(struct_array.column(1).as_ref(), int.as_ref());
assert_eq!(4, struct_array.len());
assert_eq!(0, struct_array.null_count());
assert_eq!(0, struct_array.offset());

Fields
len: usizedata_type: DataTypenulls: Option<NullBuffer>fields: Vec<Arc<dyn Array>>
Implementations

impl StructArray

pub fn new( fields: Fields, arrays: Vec<Arc<dyn Array>>, nulls: Option<NullBuffer>, ) -> StructArray

Create a new StructArray from the provided parts, panicking on failure

Panics

Panics if Self::try_new returns an error

pub fn try_new( fields: Fields, arrays: Vec<Arc<dyn Array>>, nulls: Option<NullBuffer>, ) -> Result<StructArray, ArrowError>

Create a new StructArray from the provided parts, returning an error on failure

Errors

Errors if

fields.len() != arrays.len()

fields[i].data_type() != arrays[i].data_type()

arrays[i].len() != arrays[j].len()

arrays[i].len() != nulls.len()

!fields[i].is_nullable() && !nulls.contains(arrays[i].nulls())

pub fn new_null(fields: Fields, len: usize) -> StructArray

Create a new StructArray of length len where all values are null

pub unsafe fn new_unchecked( fields: Fields, arrays: Vec<Arc<dyn Array>>, nulls: Option<NullBuffer>, ) -> StructArray

Create a new StructArray from the provided parts without validation

Safety

Safe if Self::new would not panic with the given arguments

pub fn new_empty_fields(len: usize, nulls: Option<NullBuffer>) -> StructArray

Create a new StructArray containing no fields

Panics

If len != nulls.len()

pub fn into_parts(self) -> (Fields, Vec<Arc<dyn Array>>, Option<NullBuffer>)

Deconstruct this array into its constituent parts

pub fn column(&self, pos: usize) -> &Arc<dyn Array>

Returns the field at pos.

pub fn num_columns(&self) -> usize

Return the number of fields in this struct array

pub fn columns(&self) -> &[Arc<dyn Array>]

Returns the fields of the struct array

pub fn column_names(&self) -> Vec<&str>

Return field names in this struct array

pub fn fields(&self) -> &Fields

Returns the Fields of this StructArray

pub fn column_by_name(&self, column_name: &str) -> Option<&Arc<dyn Array>>

Return child array whose field name equals to column_name

Note: A schema can currently have duplicate field names, in which case the first field will always be selected. This issue will be addressed in ARROW-11178

pub fn slice(&self, offset: usize, len: usize) -> StructArray

Returns a zero-copy slice of this array with the indicated offset and length.