arrow::array::builder

Struct FixedSizeBinaryBuilder

pub struct FixedSizeBinaryBuilder {
    values_builder: BufferBuilder<u8>,
    null_buffer_builder: NullBufferBuilder,
    value_length: i32,
}

Builder for FixedSizeBinaryArray

let mut builder = FixedSizeBinaryBuilder::with_capacity(3, 5);
// [b"hello", null, b"arrow"]
builder.append_value(b"hello").unwrap();
builder.append_null();
builder.append_value(b"arrow").unwrap();

let array = builder.finish();
assert_eq!(array.value(0), b"hello");
assert!(array.is_null(1));
assert_eq!(array.value(2), b"arrow");

Fields
values_builder: BufferBuilder<u8>null_buffer_builder: NullBufferBuildervalue_length: i32
Implementations

impl FixedSizeBinaryBuilder

pub fn new(byte_width: i32) -> FixedSizeBinaryBuilder

Creates a new FixedSizeBinaryBuilder

pub fn with_capacity(capacity: usize, byte_width: i32) -> FixedSizeBinaryBuilder

Creates a new FixedSizeBinaryBuilder, capacity is the number of byte slices that can be appended without reallocating

pub fn append_value( &mut self, value: impl AsRef<[u8]>, ) -> Result<(), ArrowError>

Appends a byte slice into the builder.

Automatically update the null buffer to delimit the slice appended in as a distinct value element.

pub fn append_null(&mut self)

Append a null value to the array.

pub fn finish(&mut self) -> FixedSizeBinaryArray

Builds the FixedSizeBinaryArray and reset this builder.

pub fn finish_cloned(&self) -> FixedSizeBinaryArray

Builds the FixedSizeBinaryArray without resetting the builder.

pub fn validity_slice(&self) -> Option<&[u8]>

Returns the current null buffer as a slice