arrow::array::builder

Struct NullBuilder

pub struct NullBuilder {
    len: usize,
}

Builder for NullArray

Example

Create a NullArray from a NullBuilder

let mut b = NullBuilder::new();
b.append_empty_value();
b.append_null();
b.append_nulls(3);
b.append_empty_values(3);
let arr = b.finish();

assert_eq!(8, arr.len());
assert_eq!(0, arr.null_count());

Fields
len: usize
Implementations

impl NullBuilder

pub fn new() -> NullBuilder

Creates a new null builder

pub fn with_capacity(_capacity: usize) -> NullBuilder

๐Deprecated: there is no actual notion of capacity in the NullBuilder, so emulating it makes little sense

Creates a new null builder with space for capacity elements without re-allocating

pub fn capacity(&self) -> usize

๐Deprecated: there is no actual notion of capacity in the NullBuilder, so emulating it makes little sense

Returns the capacity of this builder measured in slots of type T

pub fn append_null(&mut self)

Appends a null slot into the builder

pub fn append_nulls(&mut self, n: usize)

Appends n nulls into the builder.

pub fn append_empty_value(&mut self)

Appends a null slot into the builder

pub fn append_empty_values(&mut self, n: usize)

Appends n nulls into the builder.

pub fn finish(&mut self) -> NullArray

Builds the NullArray and reset this builder.

pub fn finish_cloned(&self) -> NullArray

Builds the NullArray without resetting the builder.