Type Alias arrow::array::builder::BinaryDictionaryBuilder

pub type BinaryDictionaryBuilder<K> = GenericByteDictionaryBuilder<K, GenericBinaryType<i32>>;

Builder for DictionaryArray of BinaryArray

// Create a dictionary array indexed by bytes whose values are binary.
// It can thus hold up to 256 distinct binary values.

let mut builder = BinaryDictionaryBuilder::<Int8Type>::new();

// The builder builds the dictionary value by value
builder.append(b"abc").unwrap();
builder.append_null();
builder.append(b"def").unwrap();
builder.append(b"def").unwrap();
builder.append(b"abc").unwrap();
let array = builder.finish();

assert_eq!(
  array.keys(),
  &Int8Array::from(vec![Some(0), None, Some(1), Some(1), Some(0)])
);

// Values are polymorphic and so require a downcast.
let av = array.values();
let ava: &BinaryArray = av.as_any().downcast_ref::<BinaryArray>().unwrap();

assert_eq!(ava.value(0), b"abc");
assert_eq!(ava.value(1), b"def");

Aliased Type

struct BinaryDictionaryBuilder<K> {
    state: RandomState,
    dedup: HashMap<usize, (), (), Global>,
    keys_builder: PrimitiveBuilder<K>,
    values_builder: GenericByteBuilder<GenericBinaryType<i32>>,
}

Fields
state: RandomStatededup: HashMap<usize, (), (), Global>keys_builder: PrimitiveBuilder<K>values_builder: GenericByteBuilder<GenericBinaryType<i32>>
Implementations

impl<K, T> GenericByteDictionaryBuilder<K, T>where K: ArrowDictionaryKeyType, T: ByteArrayType,

pub fn new() -> GenericByteDictionaryBuilder<K, T>

Creates a new GenericByteDictionaryBuilder

pub fn with_capacity( keys_capacity: usize, value_capacity: usize, data_capacity: usize ) -> GenericByteDictionaryBuilder<K, T>

Creates a new GenericByteDictionaryBuilder with the provided capacities

keys_capacity: the number of keys, i.e. length of array to build value_capacity: the number of distinct dictionary values, i.e. size of dictionary data_capacity: the total number of bytes of all distinct bytes in the dictionary

pub fn new_with_dictionary( keys_capacity: usize, dictionary_values: &GenericByteArray<T> ) -> Result<GenericByteDictionaryBuilder<K, T>, ArrowError>

Creates a new GenericByteDictionaryBuilder from a keys capacity and a dictionary which is initialized with the given values. The indices of those dictionary values are used as keys.

Example

let dictionary_values = StringArray::from(vec![None, Some("abc"), Some("def")]);

let mut builder = StringDictionaryBuilder::new_with_dictionary(3, &dictionary_values).unwrap();
builder.append("def").unwrap();
builder.append_null();
builder.append("abc").unwrap();

let dictionary_array = builder.finish();

let keys = dictionary_array.keys();

assert_eq!(keys, &Int16Array::from(vec![Some(2), None, Some(1)]));

impl<K, T> GenericByteDictionaryBuilder<K, T>where K: ArrowDictionaryKeyType, T: ByteArrayType,

pub fn append( &mut self, value: impl AsRef<<T as ByteArrayType>::Native> ) -> Result<<K as ArrowPrimitiveType>::Native, ArrowError>

Append a value to the array. Return an existing index if already present in the values array or a new index if the value is appended to the values array.

Returns an error if the new index would overflow the key type.

pub fn append_value(&mut self, value: impl AsRef<<T as ByteArrayType>::Native>)

Infallibly append a value to this builder

Panics

Panics if the resulting length of the dictionary values array would exceed T::Native::MAX

pub fn append_null(&mut self)

Appends a null slot into the builder

pub fn append_option( &mut self, value: Option<impl AsRef<<T as ByteArrayType>::Native>> )

Append an Option value into the builder

Panics

Panics if the resulting length of the dictionary values array would exceed T::Native::MAX

pub fn finish(&mut self) -> DictionaryArray<K>

Builds the DictionaryArray and reset this builder.

pub fn finish_cloned(&self) -> DictionaryArray<K>

Builds the DictionaryArray without resetting the builder.