Type Alias arrow::array::builder::ListBuilder

pub type ListBuilder<T> = GenericListBuilder<i32, T>;

Builder for ListArray

Aliased Type

struct ListBuilder<T> {
    offsets_builder: BufferBuilder<i32>,
    null_buffer_builder: NullBufferBuilder,
    values_builder: T,
}

Fields
offsets_builder: BufferBuilder<i32>null_buffer_builder: NullBufferBuildervalues_builder: T
Implementations

impl<OffsetSize, T> GenericListBuilder<OffsetSize, T>where OffsetSize: OffsetSizeTrait, T: ArrayBuilder,

pub fn new(values_builder: T) -> GenericListBuilder<OffsetSize, T>

Creates a new GenericListBuilder from a given values array builder

pub fn with_capacity( values_builder: T, capacity: usize ) -> GenericListBuilder<OffsetSize, T>

Creates a new GenericListBuilder from a given values array builder capacity is the number of items to pre-allocate space for in this builder

impl<OffsetSize, T> GenericListBuilder<OffsetSize, T>where OffsetSize: OffsetSizeTrait, T: ArrayBuilder + 'static,

pub fn values(&mut self) -> &mut T

Returns the child array builder as a mutable reference.

This mutable reference can be used to append values into the child array builder, but you must call append to delimit each distinct list value.

pub fn values_ref(&self) -> &T

Returns the child array builder as an immutable reference

pub fn append(&mut self, is_valid: bool)

Finish the current variable-length list array slot

Panics

Panics if the length of Self::values exceeds OffsetSize::MAX

pub fn append_value<I, V>(&mut self, i: I)where T: Extend<Option<V>>, I: IntoIterator<Item = Option<V>>,

Append a value to this GenericListBuilder

let mut builder = ListBuilder::new(Int32Builder::new());

builder.append_value([Some(1), Some(2), Some(3)]);
builder.append_value([]);
builder.append_value([None]);

let array = builder.finish();
assert_eq!(array.len(), 3);

assert_eq!(array.value_offsets(), &[0, 3, 3, 4]);
let values = array.values().as_primitive::<Int32Type>();
assert_eq!(values, &Int32Array::from(vec![Some(1), Some(2), Some(3), None]));

This is an alternative API to appending directly to Self::values and delimiting the result with Self::append

let mut builder = ListBuilder::new(Int32Builder::new());

builder.values().append_value(1);
builder.values().append_value(2);
builder.values().append_value(3);
builder.append(true);
builder.append(true);
builder.values().append_null();
builder.append(true);

let array = builder.finish();
assert_eq!(array.len(), 3);

assert_eq!(array.value_offsets(), &[0, 3, 3, 4]);
let values = array.values().as_primitive::<Int32Type>();
assert_eq!(values, &Int32Array::from(vec![Some(1), Some(2), Some(3), None]));

pub fn append_null(&mut self)

Append a null to this GenericListBuilder

See Self::append_value for an example use.

pub fn append_option<I, V>(&mut self, i: Option<I>)where T: Extend<Option<V>>, I: IntoIterator<Item = Option<V>>,

Appends an optional value into this GenericListBuilder

If Some calls Self::append_value otherwise calls Self::append_null

pub fn finish(&mut self) -> GenericListArray<OffsetSize>

Builds the GenericListArray and reset this builder.

pub fn finish_cloned(&self) -> GenericListArray<OffsetSize>

Builds the GenericListArray without resetting the builder.

pub fn offsets_slice(&self) -> &[OffsetSize]

Returns the current offsets buffer as a slice