arrow::array::ffi

Struct FFI_ArrowArray

#[repr(C)]
pub struct FFI_ArrowArray {
    length: i64,
    null_count: i64,
    offset: i64,
    n_buffers: i64,
    n_children: i64,
    buffers: *mut *const c_void,
    children: *mut *mut FFI_ArrowArray,
    dictionary: *mut FFI_ArrowArray,
    release: Option<unsafe extern "C" fn(_: *mut FFI_ArrowArray)>,
    private_data: *mut c_void,
}

ABI-compatible struct for ArrowArray from C Data Interface See https://arrow.apache.org/docs/format/CDataInterface.html#structure-definitions

fn export_array(array: &ArrayData) -> FFI_ArrowArray {
    FFI_ArrowArray::new(array)
}

Fields
length: i64null_count: i64offset: i64n_buffers: i64n_children: i64buffers: *mut *const c_voidchildren: *mut *mut FFI_ArrowArraydictionary: *mut FFI_ArrowArrayrelease: Option<unsafe extern "C" fn(_: *mut FFI_ArrowArray)>private_data: *mut c_void
Implementations

impl FFI_ArrowArray

pub fn new(data: &ArrayData) -> FFI_ArrowArray

creates a new FFI_ArrowArray from existing data.

pub unsafe fn from_raw(array: *mut FFI_ArrowArray) -> FFI_ArrowArray

Takes ownership of the pointed to FFI_ArrowArray

This acts to move the data out of array, setting the release callback to NULL

Safety

array must be valid for reads and writes

array must be properly aligned

array must point to a properly initialized value of FFI_ArrowArray

pub fn empty() -> FFI_ArrowArray

create an empty FFI_ArrowArray, which can be used to import data into

pub fn len(&self) -> usize

the length of the array

pub fn is_empty(&self) -> bool

whether the array is empty

pub fn is_released(&self) -> bool

Whether the array has been released

pub fn offset(&self) -> usize

the offset of the array

pub fn null_count(&self) -> usize

the null count of the array

pub fn null_count_opt(&self) -> Option<usize>

Returns the null count, checking for validity

pub unsafe fn set_null_count(&mut self, null_count: i64)

Set the null count of the array

Safety

Null count must match that of null buffer

pub fn buffer(&self, index: usize) -> *const u8

Returns the buffer at the provided index

Panic

Panics if index >= self.num_buffers() or the buffer is not correctly aligned

pub fn num_buffers(&self) -> usize

Returns the number of buffers

pub fn child(&self, index: usize) -> &FFI_ArrowArray

Returns the child at the provided index

pub fn num_children(&self) -> usize

Returns the number of children

pub fn dictionary(&self) -> Option<&FFI_ArrowArray>

Returns the dictionary if any