Struct arrow::array::BooleanArray

pub struct BooleanArray { /* fields omitted */ }

Array of bools

Implementations

impl BooleanArray

pub fn len(&self) -> usize

Returns the length of this array.

pub fn is_empty(&self) -> bool

Returns whether this array is empty.

pub fn builder(capacity: usize) -> BooleanBuilder

pub fn values(&self) -> &Buffer

Returns a Buffer holding all the values of this array.

Note this doesn't take the offset of this array into account.

pub fn value(&self, i: usize) -> bool

Returns the boolean value at index i.

Note this doesn't do any bound checking, for performance reason.

impl<'a> BooleanArray

pub fn iter(&'a self) -> BooleanIter<'a>

constructs a new iterator