Struct arrow::array::BufferBuilder

pub struct BufferBuilder<T: ArrowNativeType> { /* private fields */ }

Builder for creating a Buffer object.

A Buffer is the underlying data structure of Arrow’s Arrays.

For all supported types, there are type definitions for the generic version of BufferBuilder<T>, e.g. UInt8BufferBuilder.

Example:

use arrow::array::UInt8BufferBuilder;

let mut builder = UInt8BufferBuilder::new(100);
builder.append_slice(&[42, 43, 44]);
builder.append(45);
let buffer = builder.finish();

assert_eq!(unsafe { buffer.typed_data::<u8>() }, &[42, 43, 44, 45]);

Implementations

impl<T: ArrowNativeType> BufferBuilder<T>

pub fn new(capacity: usize) -> Self

Creates a new builder with initial capacity for at least capacity elements of type T.

The capacity can later be manually adjusted with the reserve() method. Also the append(), append_slice() and advance() methods automatically increase the capacity if needed.

Example:

use arrow::array::UInt8BufferBuilder;

let mut builder = UInt8BufferBuilder::new(10);

assert!(builder.capacity() >= 10);

pub fn len(&self) -> usize

Returns the current number of array elements in the internal buffer.

Example:

use arrow::array::UInt8BufferBuilder;

let mut builder = UInt8BufferBuilder::new(10);
builder.append(42);

assert_eq!(builder.len(), 1);

pub fn is_empty(&self) -> bool

Returns whether the internal buffer is empty.

Example:

use arrow::array::UInt8BufferBuilder;

let mut builder = UInt8BufferBuilder::new(10);
builder.append(42);

assert_eq!(builder.is_empty(), false);

pub fn capacity(&self) -> usize

Returns the actual capacity (number of elements) of the internal buffer.

Note: the internal capacity returned by this method might be larger than what you’d expect after setting the capacity in the new() or reserve() functions.

pub fn advance(&mut self, i: usize)

Increases the number of elements in the internal buffer by n and resizes the buffer as needed.

The values of the newly added elements are 0. This method is usually used when appending NULL values to the buffer as they still require physical memory space.

Example:

use arrow::array::UInt8BufferBuilder;

let mut builder = UInt8BufferBuilder::new(10);
builder.advance(2);

assert_eq!(builder.len(), 2);

pub fn reserve(&mut self, n: usize)

Reserves memory for at least n more elements of type T.

Example:

use arrow::array::UInt8BufferBuilder;

let mut builder = UInt8BufferBuilder::new(10);
builder.reserve(10);

assert!(builder.capacity() >= 20);

pub fn append(&mut self, v: T)

Appends a value of type T into the builder, growing the internal buffer as needed.

Example:

use arrow::array::UInt8BufferBuilder;

let mut builder = UInt8BufferBuilder::new(10);
builder.append(42);

assert_eq!(builder.len(), 1);

pub fn append_n(&mut self, n: usize, v: T)

Appends a value of type T into the builder N times, growing the internal buffer as needed.

Example:

use arrow::array::UInt8BufferBuilder;

let mut builder = UInt8BufferBuilder::new(10);
builder.append_n(10, 42);

assert_eq!(builder.len(), 10);

pub fn append_n_zeroed(&mut self, n: usize)

Appends n, zero-initialized values

Example:

use arrow::array::UInt32BufferBuilder;

let mut builder = UInt32BufferBuilder::new(10);
builder.append_n_zeroed(3);

assert_eq!(builder.len(), 3);
assert_eq!(builder.as_slice(), &[0, 0, 0])

pub fn append_slice(&mut self, slice: &[T])

Appends a slice of type T, growing the internal buffer as needed.

Example:

use arrow::array::UInt8BufferBuilder;

let mut builder = UInt8BufferBuilder::new(10);
builder.append_slice(&[42, 44, 46]);

assert_eq!(builder.len(), 3);

pub fn as_slice(&self) -> &[T]

View the contents of this buffer as a slice

use arrow::array::Float64BufferBuilder;

let mut builder = Float64BufferBuilder::new(10);
builder.append(1.3);
builder.append_n(2, 2.3);

assert_eq!(builder.as_slice(), &[1.3, 2.3, 2.3]);

pub fn as_slice_mut(&mut self) -> &mut [T]

View the contents of this buffer as a mutable slice

Example:

use arrow::array::Float32BufferBuilder;

let mut builder = Float32BufferBuilder::new(10);

builder.append_slice(&[1., 2., 3.4]);
assert_eq!(builder.as_slice(), &[1., 2., 3.4]);

builder.as_slice_mut()[1] = 4.2;
assert_eq!(builder.as_slice(), &[1., 4.2, 3.4]);

pub fn truncate(&mut self, len: usize)

Shorten this BufferBuilder to len items

If len is greater than the builder’s current length, this has no effect

Example:

use arrow::array::UInt16BufferBuilder;

let mut builder = UInt16BufferBuilder::new(10);

builder.append_slice(&[42, 44, 46]);
assert_eq!(builder.as_slice(), &[42, 44, 46]);

builder.truncate(2);
assert_eq!(builder.as_slice(), &[42, 44]);

builder.append(12);
assert_eq!(builder.as_slice(), &[42, 44, 12]);

pub unsafe fn append_trusted_len_iter(
&mut self,
iter: impl IntoIterator<Item = T>
)

Safety

This requires the iterator be a trusted length. This could instead require the iterator implement TrustedLen once that is stabilized.

pub fn finish(&mut self) -> Buffer

Resets this builder and returns an immutable Buffer.

Example:

use arrow::array::UInt8BufferBuilder;

let mut builder = UInt8BufferBuilder::new(10);
builder.append_slice(&[42, 44, 46]);

let buffer = builder.finish();

assert_eq!(unsafe { buffer.typed_data::<u8>() }, &[42, 44, 46]);