Struct arrow::array::FixedSizeListArray

pub struct FixedSizeListArray { /* fields omitted */ }

A list array where each element is a fixed-size sequence of values with the same type whose maximum length is represented by a i32.

Implementations

impl FixedSizeListArray

pub fn values(&self) -> ArrayRef

Returns a reference to the values of this list.

pub fn value_type(&self) -> DataType

Returns a clone of the value type of this list.

pub fn value(&self, i: usize) -> ArrayRef

Returns ith value of this list array.

pub fn value_offset(&self, i: usize) -> i32

Returns the offset for value at index i.

Note this doesn't do any bound checking, for performance reason.

pub const fn value_length(&self) -> i32

Returns the length for value at index i.

Note this doesn't do any bound checking, for performance reason.