Struct arrow::array::GenericBinaryArray

pub struct GenericBinaryArray<OffsetSize: OffsetSizeTrait> { /* private fields */ }

See BinaryArray and LargeBinaryArray for storing binary data.

Implementations

impl<OffsetSize: OffsetSizeTrait> GenericBinaryArray<OffsetSize>

pub fn get_data_type() -> DataType

Get the data type of the array.

pub fn value_length(&self, i: usize) -> OffsetSize

Returns the length for value at index i.

pub fn value_data(&self) -> Buffer

Returns a clone of the value data buffer

pub fn value_offsets(&self) -> &[OffsetSize]

Returns the offset values in the offsets buffer

pub unsafe fn value_unchecked(&self, i: usize) -> &[u8]

Returns the element at index i as bytes slice

Safety

Caller is responsible for ensuring that the index is within the bounds of the array

pub fn value(&self, i: usize) -> &[u8]

Returns the element at index i as bytes slice

pub fn from_vec(v: Vec<&[u8]>) -> Self

Creates a GenericBinaryArray from a vector of byte slices

See also Self::from_iter_values

pub fn from_opt_vec(v: Vec<Option<&[u8]>>) -> Self

Creates a GenericBinaryArray from a vector of Optional (null) byte slices

pub fn from_iter_values<Ptr, I>(iter: I) -> Self where
Ptr: AsRef<[u8]>,
I: IntoIterator<Item = Ptr>,

Creates a GenericBinaryArray based on an iterator of values without nulls

pub fn take_iter<'a>(
&'a self,
indexes: impl Iterator<Item = Option<usize>> + 'a
) -> impl Iterator<Item = Option<&[u8]>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

pub unsafe fn take_iter_unchecked<'a>(
&'a self,
indexes: impl Iterator<Item = Option<usize>> + 'a
) -> impl Iterator<Item = Option<&[u8]>> + 'a

Returns an iterator that returns the values of array.value(i) for an iterator with each element i

Safety

caller must ensure that the indexes in the iterator are less than the array.len()

impl<'a, T: OffsetSizeTrait> GenericBinaryArray<T>

pub fn iter(&'a self) -> GenericBinaryIter<'a, T>

constructs a new iterator