Struct arrow::array::GenericListArray

pub struct GenericListArray<OffsetSize> { /* fields omitted */ }

Generic struct for a primitive Array

For non generic lists, you may wish to consider using ListArray or LargeListArray`

Implementations

impl<OffsetSize: OffsetSizeTrait> GenericListArray<OffsetSize>

pub fn values(&self) -> ArrayRef

Returns a reference to the values of this list.

pub fn value_type(&self) -> DataType

Returns a clone of the value type of this list.

pub unsafe fn value_unchecked(&self, i: usize) -> ArrayRef

Returns ith value of this list array.

Safety

Caller must ensure that the index is within the array bounds

pub fn value(&self, i: usize) -> ArrayRef

Returns ith value of this list array.

pub fn value_offsets(&self) -> &[OffsetSize]

Returns the offset values in the offsets buffer

pub fn value_length(&self, i: usize) -> OffsetSize

Returns the length for value at index i.

pub fn iter<'a>(&'a self) -> GenericListArrayIter<'a, OffsetSize>

constructs a new iterator

pub fn from_iter_primitive<T, P, I>(iter: I) -> Self where
T: ArrowPrimitiveType,
P: AsRef<[Option<<T as ArrowPrimitiveType>::Native>]> + IntoIterator<Item = Option<<T as ArrowPrimitiveType>::Native>>,
I: IntoIterator<Item = Option<P>>,

Creates a GenericListArray from an iterator of primitive values

Example

let data = vec![
   Some(vec![Some(0), Some(1), Some(2)]),
   None,
   Some(vec![Some(3), None, Some(5)]),
   Some(vec![Some(6), Some(7)]),
];
let list_array = ListArray::from_iter_primitive::<Int32Type, _, _>(data);
println!("{:?}", list_array);