Struct arrow::array::MapArray

pub struct MapArray { /* private fields */ }

A nested array type where each record is a key-value map. Keys should always be non-null, but values can be null.

MapArray is physically a crate::array::ListArray that has a crate::array::StructArray with 2 child fields.

Implementations

impl MapArray

pub fn keys(&self) -> ArrayRef

Returns a reference to the keys of this map.

pub fn values(&self) -> ArrayRef

Returns a reference to the values of this map.

pub fn key_type(&self) -> DataType

Returns the data type of the map’s keys.

pub fn value_type(&self) -> DataType

Returns the data type of the map’s values.

pub unsafe fn value_unchecked(&self, i: usize) -> ArrayRef

Returns ith value of this map array.

Safety

Caller must ensure that the index is within the array bounds

pub fn value(&self, i: usize) -> ArrayRef

Returns ith value of this map array.

pub fn value_offsets(&self) -> &[i32]

Returns the offset values in the offsets buffer

pub fn value_length(&self, i: usize) -> i32

Returns the length for value at index i.

impl MapArray

pub fn new_from_strings<'a>(
keys: impl Iterator<Item = &'a str>,
values: &dyn Array,
entry_offsets: &[u32]
) -> Result<Self, ArrowError>

Creates map array from provided keys, values and entry_offsets.