Struct arrow::array::StructArray

pub struct StructArray { /* fields omitted */ }

A nested array type where each child (called field) is represented by a separate array.

Implementations

impl StructArray

pub fn column(&self, pos: usize) -> &ArrayRef

Returns the field at pos.

pub fn num_columns(&self) -> usize

Return the number of fields in this struct array

pub fn columns(&self) -> Vec<&ArrayRef>

Returns the fields of the struct array

pub fn columns_ref(&self) -> Vec<ArrayRef>

Returns child array refs of the struct array

pub fn column_names(&self) -> Vec<&str>

Return field names in this struct array

pub fn column_by_name(&self, column_name: &str) -> Option<&ArrayRef>

Return child array whose field name equals to column_name