Module arrow::buffer

This module contains two main structs: Buffer and MutableBuffer. A buffer represents a contiguous memory region that can be shared via offsets.

Structs

Buffer

Buffer represents a contiguous memory region that can be shared with other buffers and across thread boundaries.

MutableBuffer

A MutableBuffer is Arrow’s interface to build a Buffer out of items or slices of items. Buffers created from MutableBuffer (via into) are guaranteed to have its pointer aligned along cache lines and in multiple of 64 bytes. Use MutableBuffer::push to insert an item, MutableBuffer::extend_from_slice to insert many items, and into to convert it to Buffer.

ScalarBuffer

Provides a safe API for interpreting a Buffer as a slice of ArrowNativeType

Functions

bitwise_bin_op_helper

Apply a bitwise operation op to two inputs and return the result as a Buffer. The inputs are treated as bitmaps, meaning that offsets and length are specified in number of bits.

bitwise_unary_op_helper

Apply a bitwise operation op to one input and return the result as a Buffer. The input is treated as a bitmap, meaning that offset and length are specified in number of bits.

buffer_bin_and

buffer_bin_or

buffer_unary_not