Struct arrow::buffer::MutableBuffer

pub struct MutableBuffer { /* fields omitted */ }

A MutableBuffer is Arrow's interface to build a Buffer out of items or slices of items. Buffers created from MutableBuffer (via into) are guaranteed to have its pointer aligned along cache lines and in multiple of 64 bytes. Use MutableBuffer::push to insert an item, MutableBuffer::extend_from_slice to insert many items, and into to convert it to Buffer.

Example

let mut buffer = MutableBuffer::new(0);
buffer.push(256u32);
buffer.extend_from_slice(&[1u32]);
let buffer: Buffer = buffer.into();
assert_eq!(buffer.as_slice(), &[0u8, 1, 0, 0, 1, 0, 0, 0])

Implementations

impl MutableBuffer

pub fn new(capacity: usize) -> Self

Allocate a new MutableBuffer with initial capacity to be at least capacity.

pub fn from_len_zeroed(len: usize) -> Self

Allocates a new MutableBuffer with len and capacity to be at least len where all bytes are guaranteed to be 0u8.

Example

let mut buffer = MutableBuffer::from_len_zeroed(127);
assert_eq!(buffer.len(), 127);
assert!(buffer.capacity() >= 127);
let data = buffer.as_slice_mut();
assert_eq!(data[126], 0u8);

pub fn new_null(len: usize) -> Self

creates a new MutableBuffer with capacity and length capable of holding len bits. This is useful to create a buffer for packed bitmaps.

pub fn with_bitset(self, end: usize, val: bool) -> Self

Set the bits in the range of [0, end) to 0 (if val is false), or 1 (if val is true). Also extend the length of this buffer to be end.

This is useful when one wants to clear (or set) the bits and then manipulate the buffer directly (e.g., modifying the buffer by holding a mutable reference from data_mut()).

pub fn set_null_bits(&mut self, start: usize, count: usize)

Ensure that count bytes from start contain zero bits

This is used to initialize the bits in a buffer, however, it has no impact on the len of the buffer and so can be used to initialize the memory region from len to capacity.

pub fn reserve(&mut self, additional: usize)

Ensures that this buffer has at least self.len + additional bytes. This re-allocates iff self.len + additional > capacity.

Example

let mut buffer = MutableBuffer::new(0);
buffer.reserve(253); // allocates for the first time
(0..253u8).for_each(|i| buffer.push(i)); // no reallocation
let buffer: Buffer = buffer.into();
assert_eq!(buffer.len(), 253);

pub fn resize(&mut self, new_len: usize, value: u8)

Resizes the buffer, either truncating its contents (with no change in capacity), or growing it (potentially reallocating it) and writing value in the newly available bytes.

Example

let mut buffer = MutableBuffer::new(0);
buffer.resize(253, 2); // allocates for the first time
assert_eq!(buffer.as_slice()[252], 2u8);

pub const fn is_empty(&self) -> bool

Returns whether this buffer is empty or not.

pub const fn len(&self) -> usize

Returns the length (the number of bytes written) in this buffer. The invariant buffer.len() <= buffer.capacity() is always upheld.

pub const fn capacity(&self) -> usize

Returns the total capacity in this buffer. The invariant buffer.len() <= buffer.capacity() is always upheld.

pub fn clear(&mut self)

Clear all existing data from this buffer.

pub fn as_slice(&self) -> &[u8]

Returns the data stored in this buffer as a slice.

pub fn as_slice_mut(&mut self) -> &mut [u8]

Returns the data stored in this buffer as a mutable slice.

pub const fn as_ptr(&self) -> *const u8

Returns a raw pointer to this buffer's internal memory This pointer is guaranteed to be aligned along cache-lines.

pub fn as_mut_ptr(&mut self) -> *mut u8

Returns a mutable raw pointer to this buffer's internal memory This pointer is guaranteed to be aligned along cache-lines.

pub fn freeze(self) -> Buffer

👎 Deprecated since 2.0.0:
This method is deprecated in favour of into from the trait Into.

Freezes this buffer and return an immutable version of it.

pub fn typed_data_mut<T: ArrowNativeType>(&mut self) -> &mut [T]

View this buffer asa slice of a specific type.

Safety

This function must only be used when this buffer was extended with items of type T. Failure to do so results in undefined behavior.

pub fn extend_from_slice<T: ToByteSlice>(&mut self, items: &[T])

Extends this buffer from a slice of items that can be represented in bytes, increasing its capacity if needed.

Example

let mut buffer = MutableBuffer::new(0);
buffer.extend_from_slice(&[2u32, 0]);
assert_eq!(buffer.len(), 8) // u32 has 4 bytes

pub fn push<T: ToByteSlice>(&mut self, item: T)

Extends the buffer with a new item, increasing its capacity if needed.

Example

let mut buffer = MutableBuffer::new(0);
buffer.push(256u32);
assert_eq!(buffer.len(), 4) // u32 has 4 bytes

pub fn extend_zeros(&mut self, additional: usize)

Extends the buffer by additional bytes equal to 0u8, incrementing its capacity if needed.