Struct arrow::bytes::Bytes

pub struct Bytes { /* fields omitted */ }

A continuous, fixed-size, immutable memory region that knows how to de-allocate itself. This structs' API is inspired by the bytes::Bytes, but it is not limited to using rust's global allocator nor u8 aligmnent.

In the most common case, this buffer is allocated using allocate_aligned and deallocated accordingly free_aligned. When the region is allocated by an foreign allocator, Deallocation::Foreign, this calls the foreign deallocator to deallocate the region when it is no longer needed.

Implementations

impl Bytes

pub unsafe fn new(
ptr: NonNull<u8>,
len: usize,
deallocation: Deallocation
) -> Bytes

Takes ownership of an allocated memory region,

Arguments

ptr - Pointer to raw parts

len - Length of raw parts in bytes

capacity - Total allocated memory for the pointer ptr, in bytes

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len bytes. If the ptr and capacity come from a Buffer, then this is guaranteed.

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn ptr(&self) -> NonNull<u8>

pub fn capacity(&self) -> usize