arrow::compute

Function max_array

pub fn max_array<T, A>(array: A) -> Option<<T as ArrowPrimitiveType>::Native>
where
    A: ArrayAccessor<Item = <T as ArrowPrimitiveType>::Native>,
    T: ArrowNumericType,
    <T as ArrowPrimitiveType>::Native: ArrowNativeTypeOp,

Returns the max of values in the array of ArrowNumericType type, or dictionary array with value of ArrowNumericType type.