arrow::compute
Function year_dyn

pub fn year_dyn(array: &dyn Array) -> Result<Arc<dyn Array>, ArrowError>

👎Deprecated since 51.0.0: Use date_part instead

Extracts the years of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.