Module arrow::compute::kernels::boolean

Defines boolean kernels on Arrow BooleanArray’s, e.g. AND, OR and NOT.

These kernels can leverage SIMD if available on your system. Currently no runtime detection is provided, you should enable the specific SIMD intrinsics using RUSTFLAGS="-C target-feature=+avx2" for example. See the documentation here for more information.

Functions

and

Performs AND operation on two arrays. If either left or right value is null then the result is also null.

and_kleene

Logical ‘and’ boolean values with Kleene logic

is_not_null

Returns a non-null BooleanArray with whether each value of the array is not null.

is_null

Returns a non-null BooleanArray with whether each value of the array is null.

not

Performs unary NOT operation on an arrays. If value is null then the result is also null.

nullif

Copies original array, setting null bit to true if a secondary comparison boolean array is set to true. Typically used to implement NULLIF.

or

Performs OR operation on two arrays. If either left or right value is null then the result is also null.

or_kleene

Logical ‘or’ boolean values with Kleene logic