Function arrow::compute::kernels::boolean::or

pub fn or(left: &BooleanArray, right: &BooleanArray) -> Result<BooleanArray>

Performs OR operation on two arrays. If either left or right value is null then the result is also null.

Error

This function errors when the arrays have different lengths.

Example

use arrow::array::BooleanArray;
use arrow::error::Result;
use arrow::compute::kernels::boolean::or;
let a = BooleanArray::from(vec![Some(false), Some(true), None]);
let b = BooleanArray::from(vec![Some(true), Some(true), Some(false)]);
let or_ab = or(&a, &b)?;
assert_eq!(or_ab, BooleanArray::from(vec![Some(true), Some(true), None]));