Module arrow::compute::kernels::cast

Defines cast kernels for ArrayRef, to convert Arrays between supported datatypes.

Example:

use arrow::array::*;
use arrow::compute::cast;
use arrow::datatypes::DataType;
use std::sync::Arc;

let a = Int32Array::from(vec![5, 6, 7]);
let array = Arc::new(a) as ArrayRef;
let b = cast(&array, &DataType::Float64).unwrap();
let c = b.as_any().downcast_ref::<Float64Array>().unwrap();
assert_eq!(5.0, c.value(0));
assert_eq!(6.0, c.value(1));
assert_eq!(7.0, c.value(2));

Functions

can_cast_types
Return true if a value of type from_type can be cast into a value of to_type. Note that such as cast may be lossy.

cast
Cast array to the provided data type and return a new Array with type to_type, if possible.