Module arrow::compute::kernels::comparison

Comparison kernels for Arrays.

These kernels can leverage SIMD if available on your system. Currently no runtime detection is provided, you should enable the specific SIMD intrinsics using RUSTFLAGS="-C target-feature=+avx2" for example. See the documentation here for more information.

Functions

cmp_dict

Perform given operation on two DictionaryArrays. Returns an error if the two arrays have different value type

cmp_dict_binary

Perform the given operation on two DictionaryArrays which value type is DataType::Binary or DataType::LargeBinary.

cmp_dict_bool

Perform the given operation on two DictionaryArrays which value type is DataType::Boolean.

cmp_dict_utf8

Perform the given operation on two DictionaryArrays which value type is DataType::Utf8 or DataType::LargeUtf8.

contains

Checks if a GenericListArray contains a value in the PrimitiveArray

contains_utf8

Checks if a GenericListArray contains a value in the GenericStringArray

eq

Perform left == right operation on two PrimitiveArrays.

eq_binary

Perform left == right operation on BinaryArray / LargeBinaryArray.

eq_binary_scalar

Perform left == right operation on BinaryArray / LargeBinaryArray and a scalar

eq_bool

Perform left == right operation on BooleanArray

eq_bool_scalar

Perform left == right operation on BooleanArray and a scalar

eq_dyn

Perform left == right operation on two (dynamic) Arrays.

eq_dyn_binary_scalar

Perform left == right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

eq_dyn_bool_scalar

Perform left == right operation on an array and a numeric scalar value.

eq_dyn_scalar

Perform left == right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values

eq_dyn_utf8_scalar

Perform left == right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

eq_scalar

Perform left == right operation on a PrimitiveArray and a scalar value.

eq_utf8

Perform left == right operation on StringArray / LargeStringArray.

eq_utf8_scalar

Perform left == right operation on StringArray / LargeStringArray and a scalar.

gt

Perform left > right operation on two PrimitiveArrays. Non-null values are greater than null values.

gt_binary

Perform left > right operation on BinaryArray / LargeBinaryArray.

gt_binary_scalar

Perform left > right operation on BinaryArray / LargeBinaryArray and a scalar.

gt_bool

Perform left > right operation on BooleanArray

gt_bool_scalar

Perform left > right operation on BooleanArray and a scalar

gt_dyn

Perform left > right operation on two (dynamic) Arrays.

gt_dyn_binary_scalar

Perform left > right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

gt_dyn_bool_scalar

Perform left > right operation on an array and a numeric scalar value. Supports BooleanArrays.

gt_dyn_scalar

Perform left > right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values

gt_dyn_utf8_scalar

Perform left > right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

gt_eq

Perform left >= right operation on two PrimitiveArrays. Non-null values are greater than null values.

gt_eq_binary

Perform left >= right operation on BinaryArray / LargeBinaryArray.

gt_eq_binary_scalar

Perform left >= right operation on BinaryArray / LargeBinaryArray and a scalar.

gt_eq_bool

Perform left >= right operation on BooleanArray

gt_eq_bool_scalar

Perform left >= right operation on BooleanArray and a scalar

gt_eq_dyn

Perform left >= right operation on two (dynamic) Arrays.

gt_eq_dyn_binary_scalar

Perform left >= right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

gt_eq_dyn_bool_scalar

Perform left >= right operation on an array and a numeric scalar value. Supports BooleanArrays.

gt_eq_dyn_scalar

Perform left >= right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values

gt_eq_dyn_utf8_scalar

Perform left >= right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

gt_eq_scalar

Perform left >= right operation on a PrimitiveArray and a scalar value. Non-null values are greater than null values.

gt_eq_utf8

Perform left >= right operation on StringArray / LargeStringArray.

gt_eq_utf8_scalar

Perform left >= right operation on StringArray / LargeStringArray and a scalar.

gt_scalar

Perform left > right operation on a PrimitiveArray and a scalar value. Non-null values are greater than null values.

gt_utf8

Perform left > right operation on StringArray / LargeStringArray.

gt_utf8_scalar

Perform left > right operation on StringArray / LargeStringArray and a scalar.

ilike_utf8

Perform SQL left ILIKE right operation on StringArray / LargeStringArray.

ilike_utf8_scalar

Perform SQL left ILIKE right operation on StringArray / LargeStringArray and a scalar.

like_utf8

Perform SQL left LIKE right operation on StringArray / LargeStringArray.

like_utf8_scalar

Perform SQL left LIKE right operation on StringArray / LargeStringArray and a scalar.

lt

Perform left < right operation on two PrimitiveArrays. Null values are less than non-null values.

lt_binary

Perform left < right operation on BinaryArray / LargeBinaryArray.

lt_binary_scalar

Perform left < right operation on BinaryArray / LargeBinaryArray and a scalar.

lt_bool

Perform left < right operation on BooleanArray

lt_bool_scalar

Perform left < right operation on BooleanArray and a scalar

lt_dyn

Perform left < right operation on two (dynamic) Arrays.

lt_dyn_binary_scalar

Perform left < right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

lt_dyn_bool_scalar

Perform left < right operation on an array and a numeric scalar value. Supports BooleanArrays.

lt_dyn_scalar

Perform left < right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values

lt_dyn_utf8_scalar

Perform left < right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

lt_eq

Perform left <= right operation on two PrimitiveArrays. Null values are less than non-null values.

lt_eq_binary

Perform left <= right operation on BinaryArray / LargeBinaryArray.

lt_eq_binary_scalar

Perform left <= right operation on BinaryArray / LargeBinaryArray and a scalar.

lt_eq_bool

Perform left <= right operation on BooleanArray

lt_eq_bool_scalar

Perform left <= right operation on BooleanArray and a scalar

lt_eq_dyn

Perform left <= right operation on two (dynamic) Arrays.

lt_eq_dyn_binary_scalar

Perform left <= right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

lt_eq_dyn_bool_scalar

Perform left <= right operation on an array and a numeric scalar value. Supports BooleanArrays.

lt_eq_dyn_scalar

Perform left <= right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values

lt_eq_dyn_utf8_scalar

Perform left <= right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

lt_eq_scalar

Perform left <= right operation on a PrimitiveArray and a scalar value. Null values are less than non-null values.

lt_eq_utf8

Perform left <= right operation on StringArray / LargeStringArray.

lt_eq_utf8_scalar

Perform left <= right operation on StringArray / LargeStringArray and a scalar.

lt_scalar

Perform left < right operation on a PrimitiveArray and a scalar value. Null values are less than non-null values.

lt_utf8

Perform left < right operation on StringArray / LargeStringArray.

lt_utf8_scalar

Perform left < right operation on StringArray / LargeStringArray and a scalar.

neq

Perform left != right operation on two PrimitiveArrays.

neq_binary

Perform left != right operation on BinaryArray / LargeBinaryArray.

neq_binary_scalar

Perform left != right operation on BinaryArray / LargeBinaryArray and a scalar.

neq_bool

Perform left != right operation on BooleanArray

neq_bool_scalar

Perform left != right operation on BooleanArray and a scalar

neq_dyn

Perform left != right operation on two (dynamic) Arrays.

neq_dyn_binary_scalar

Perform left != right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

neq_dyn_bool_scalar

Perform left != right operation on an array and a numeric scalar value. Supports BooleanArrays.

neq_dyn_scalar

Perform left != right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values

neq_dyn_utf8_scalar

Perform left != right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

neq_scalar

Perform left != right operation on a PrimitiveArray and a scalar value.

neq_utf8

Perform left != right operation on StringArray / LargeStringArray.

neq_utf8_scalar

Perform left != right operation on StringArray / LargeStringArray and a scalar.

nilike_utf8

Perform SQL left NOT ILIKE right operation on StringArray / LargeStringArray.

nilike_utf8_scalar

Perform SQL left NOT ILIKE right operation on StringArray / LargeStringArray and a scalar.

nlike_utf8

Perform SQL left NOT LIKE right operation on StringArray / LargeStringArray.

nlike_utf8_scalar

Perform SQL left NOT LIKE right operation on StringArray / LargeStringArray and a scalar.

no_simd_compare_op

Evaluate op(left, right) for PrimitiveArrays using a specified comparison function.

no_simd_compare_op_scalar

Evaluate op(left, right) for PrimitiveArray and scalar using a specified comparison function.

regexp_is_match_utf8

Perform SQL array ~ regex_array operation on StringArray / LargeStringArray. If regex_array element has an empty value, the corresponding result value is always true.

regexp_is_match_utf8_scalar

Perform SQL array ~ regex_array operation on StringArray / LargeStringArray and a scalar.