Function arrow::compute::kernels::comparison::eq_dyn

pub fn eq_dyn(
    left: &dyn Array,
    right: &dyn Array
) -> Result<BooleanArray, ArrowError>

👎Deprecated: Use arrow_ord::cmp::eq

Perform left == right operation on two (dynamic) Arrays.

Only when two arrays are of the same type the comparison will happen otherwise it will err with a casting error.

For floating values like f32 and f64, this comparison produces an ordering in accordance to the totalOrder predicate as defined in the IEEE 754 (2008 revision) floating point standard. Note that totalOrder treats positive and negative zeros are different. If it is necessary to treat them as equal, please normalize zeros before calling this kernel. Please refer to f32::total_cmp and f64::total_cmp.

Example

use arrow_array::{StringArray, BooleanArray};
use arrow_ord::comparison::eq_dyn;
let array1 = StringArray::from(vec![Some("foo"), None, Some("bar")]);
let array2 = StringArray::from(vec![Some("foo"), None, Some("baz")]);
let result = eq_dyn(&array1, &array2).unwrap();
assert_eq!(BooleanArray::from(vec![Some(true), None, Some(false)]), result);