Module arrow::compute::kernels::filter

Defines miscellaneous array kernels.

Structs

FilterBuilder

A builder to construct FilterPredicate

FilterPredicate

A filtering predicate that can be applied to an Array

SlicesIterator

An iterator of (usize, usize) each representing an interval [start, end) whose slots of a BooleanArray are true. Each interval corresponds to a contiguous region of memory to be “taken” from an array to be filtered.

Functions

build_filterDeprecated

Returns a prepared function optimized to filter multiple arrays. Creating this function requires time, but using it is faster than filter when the same filter needs to be applied to multiple arrays (e.g. a multi-column RecordBatch). WARNING: the nulls of filter are ignored and the value on its slot is considered. Therefore, it is considered undefined behavior to pass filter with null values.

filter

Filters an Array, returning elements matching the filter (i.e. where the values are true).

filter_record_batch

Returns a new RecordBatch with arrays containing only values matching the filter.

prep_null_mask_filter

Remove null values by do a bitmask AND operation with null bits and the boolean bits.

Type Definitions

FilterDeprecated

Function that can filter arbitrary arrays