Function arrow::compute::kernels::filter::filter

pub fn filter(array: &dyn Array, filter: &BooleanArray) -> Result<ArrayRef>

Filters an Array, returning elements matching the filter (i.e. where the values are true). WARNING: the nulls of filter are ignored and the value on its slot is considered. Therefore, it is considered undefined behavior to pass filter with null values.

Example

let array = Int32Array::from(vec![5, 6, 7, 8, 9]);
let filter_array = BooleanArray::from(vec![true, false, false, true, false]);
let c = filter(&array, &filter_array)?;
let c = c.as_any().downcast_ref::<Int32Array>().unwrap();
assert_eq!(c, &Int32Array::from(vec![5, 8]));