arrow::compute::kernels

Module numeric

Defines numeric arithmetic kernels on PrimitiveArray, such as add

Functions

add

Perform lhs + rhs, returning an error on overflow

add_wrapping

Perform lhs + rhs, wrapping on overflow for DataType::is_integer

div

Perform lhs / rhs

mul

Perform lhs * rhs, returning an error on overflow

mul_wrapping

Perform lhs * rhs, wrapping on overflow for DataType::is_integer

neg

Negates each element of array, returning an error on overflow

neg_wrapping

Negates each element of array, wrapping on overflow for DataType::is_integer

rem

Perform lhs % rhs

sub

Perform lhs - rhs, returning an error on overflow

sub_wrapping

Perform lhs - rhs, wrapping on overflow for DataType::is_integer