Module arrow::compute::kernels::regexp

Defines kernel to extract substrings based on a regular expression of a [Large]StringArray

Functions

regexp_is_match_utf8

Perform SQL array ~ regex_array operation on StringArray / LargeStringArray. If regex_array element has an empty value, the corresponding result value is always true.

regexp_is_match_utf8_scalar

Perform SQL array ~ regex_array operation on StringArray / LargeStringArray and a scalar.

regexp_match

Extract all groups matched by a regular expression for a given String array.