Function arrow::compute::kernels::regexp::regexp_is_match_utf8

pub fn regexp_is_match_utf8<OffsetSize>(
    array: &GenericByteArray<GenericStringType<OffsetSize>>,
    regex_array: &GenericByteArray<GenericStringType<OffsetSize>>,
    flags_array: Option<&GenericByteArray<GenericStringType<OffsetSize>>>
) -> Result<BooleanArray, ArrowError>
where
    OffsetSize: OffsetSizeTrait,

Perform SQL array ~ regex_array operation on StringArray / LargeStringArray. If regex_array element has an empty value, the corresponding result value is always true.

flags_array are optional StringArray / LargeStringArray flag, which allow special search modes, such as case insensitive and multi-line mode. See the documentation here for more information.