Module arrow::compute::kernels::sort

Defines sort kernel for ArrayRef

Structs

SortColumn

One column to be used in lexicographical sort

SortOptions

Options that define how sort kernels should behave

Functions

lexsort

Sort a list of ArrayRef using SortOptions provided for each array.

lexsort_to_indices

Sort elements lexicographically from a list of ArrayRef into an unsigned integer (UInt32Array) of indices.

partial_sort

It’s unstable_sort, may not preserve the order of equal elements

sort

Sort the ArrayRef using SortOptions.

sort_limit

Sort the ArrayRef partially.

sort_to_indices

Sort elements from ArrayRef into an unsigned integer (UInt32Array) of indices. For floating point arrays any NaN values are considered to be greater than any other non-null value limit is an option for partial_sort