arrow::compute::kernels::sort

Struct LexicographicalComparator

pub struct LexicographicalComparator {
    compare_items: Vec<Box<dyn Fn(usize, usize) -> Ordering + Sync + Send>>,
}

A lexicographical comparator that wraps given array data (columns) and can lexicographically compare data at given two indices. The lifetime is the same at the data wrapped.

Fields
compare_items: Vec<Box<dyn Fn(usize, usize) -> Ordering + Sync + Send>>
Implementations

impl LexicographicalComparator

pub fn compare(&self, a_idx: usize, b_idx: usize) -> Ordering

lexicographically compare values at the wrapped columns with given indices.

pub fn try_new( columns: &[SortColumn], ) -> Result<LexicographicalComparator, ArrowError>

Create a new lex comparator that will wrap the given sort columns and give comparison results with two indices.