Module arrow::compute::kernels::substring

Defines kernel to extract a substring of an Array Supported array types: GenericStringArray, GenericBinaryArray, FixedSizeBinaryArray, DictionaryArray

Functions

substring

Returns an ArrayRef with substrings of all the elements in array.

substring_by_char

Substrings based on character index