arrow::compute::kernels

Module temporal

Defines temporal kernels for time and date related functions.

Enums

DatePart

Valid parts to extract from date/time/timestamp arrays.

Functions

date_part

Given an array, return a new array with the extracted DatePart as signed 32-bit integer values.