Struct arrow::csv::reader::ReaderBuilder

pub struct ReaderBuilder { /* private fields */ }

CSV file reader builder

Implementations

impl ReaderBuilder

pub fn new() -> ReaderBuilder

Create a new builder for configuring CSV parsing options.

To convert a builder into a reader, call ReaderBuilder::build

Example

extern crate arrow;

use arrow::csv;
use std::fs::File;

fn example() -> csv::Reader<File> {
    let file = File::open("test/data/uk_cities_with_headers.csv").unwrap();

    // create a builder, inferring the schema with the first 100 records
    let builder = csv::ReaderBuilder::new().infer_schema(Some(100));

    let reader = builder.build(file).unwrap();

    reader
}

pub fn with_schema(self, schema: SchemaRef) -> Self

Set the CSV file’s schema

pub fn has_header(self, has_header: bool) -> Self

Set whether the CSV file has headers

pub fn with_datetime_re(self, datetime_re: Regex) -> Self

Set the datetime regex used to parse the string to Date64Type this regex is used while infering schema

pub fn with_datetime_format(self, datetime_format: String) -> Self

Set the datetime fromat used to parse the string to Date64Type this fromat is used while when the schema wants to parse Date64Type.

For format refer to chrono docs

pub fn with_delimiter(self, delimiter: u8) -> Self

Set the CSV file’s column delimiter as a byte character

pub fn with_escape(self, escape: u8) -> Self

pub fn with_quote(self, quote: u8) -> Self

pub fn with_terminator(self, terminator: u8) -> Self

pub fn infer_schema(self, max_records: Option<usize>) -> Self

Set the CSV reader to infer the schema of the file

pub fn with_batch_size(self, batch_size: usize) -> Self

Set the batch size (number of records to load at one time)

pub fn with_bounds(self, start: usize, end: usize) -> Self

Set the bounds over which to scan the reader. start and end are line numbers.

pub fn with_projection(self, projection: Vec<usize>) -> Self

Set the reader’s column projection

pub fn build<R: Read + Seek>(self, reader: R) -> Result<Reader<R>>

Create a new Reader from the ReaderBuilder