Module arrow::datatypes

Defines the logical data types of Arrow arrays.

The most important things you might be looking for are:

Schema to describe a schema.

Field to describe one field within a schema.

DataType to describe the type of a field.

Structs

BooleanType

Date32Type

Date64Type

DurationMicrosecondType

DurationMillisecondType

DurationNanosecondType

DurationSecondType

Field

Contains the meta-data for a single relative type.

Float16Type

Float32Type

Float64Type

Int8Type

Int16Type

Int32Type

Int64Type

IntervalDayTimeType

IntervalMonthDayNanoType

IntervalYearMonthType

Schema

Describes the meta-data of an ordered sequence of relative types.

Time32MillisecondType

Time32SecondType

Time64MicrosecondType

Time64NanosecondType

TimestampMicrosecondType

TimestampMillisecondType

TimestampNanosecondType

TimestampSecondType

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Enums

DataType

The set of datatypes that are supported by this implementation of Apache Arrow.

IntervalUnit

YEAR_MONTH, DAY_TIME, MONTH_DAY_NANO interval in SQL style.

TimeUnit

An absolute length of time in seconds, milliseconds, microseconds or nanoseconds.

UnionMode

Constants

DECIMAL_DEFAULT_SCALE

The default scale for DataType::Decimal values

DECIMAL_MAX_PRECISION

The maximum precision for DataType::Decimal values

DECIMAL_MAX_SCALE

The maximum scale for DataType::Decimal values

MAX_DECIMAL_FOR_EACH_PRECISION

MAX_DECIMAL_FOR_EACH_PRECISION[p] holds the maximum i128 value that can be stored in DataType::Decimal value of precision p

MIN_DECIMAL_FOR_EACH_PRECISION

MIN_DECIMAL_FOR_EACH_PRECISION[p] holds the minimum i128 value that can be stored in a DataType::Decimal value of precision p

Traits

ArrowDictionaryKeyType

A subtype of primitive type that represents legal dictionary keys. See https://arrow.apache.org/docs/format/Columnar.html

ArrowFloatNumericType

ArrowNativeType

Trait expressing a Rust type that has the same in-memory representation as Arrow. This includes i16, f32, but excludes bool (which in arrow is represented in bits).

ArrowNumericType

ArrowPrimitiveType

Trait bridging the dynamic-typed nature of Arrow (via DataType) with the static-typed nature of rust types (ArrowNativeType) for all types that implement ArrowNativeType.

ArrowTemporalType

A subtype of primitive type that represents temporal values.

ArrowTimestampType

A timestamp type allows us to create array builders that take a timestamp.

JsonSerializable

Trait declaring any type that is serializable to JSON. This includes all primitive types (bool, i32, etc.).

ToByteSlice

Allows conversion from supported Arrow types to a byte slice.

Type Definitions

SchemaRef

A reference-counted reference to a Schema.