Struct arrow::datatypes::Field

pub struct Field { /* private fields */ }

Describes a single column in a Schema.

A Schema is an ordered collection of Field objects.

Implementations

impl Field

pub fn new( name: impl Into<String>, data_type: DataType, nullable: bool ) -> Field

Creates a new field with the given name, type, and nullability

pub fn new_list_field(data_type: DataType, nullable: bool) -> Field

Creates a new Field`` suitable for [DataType::List] and [DataType::LargeList`]

While not required, this method follows the convention of naming the Field "item".

Example

assert_eq!(
  Field::new("item", DataType::Int32, true),
  Field::new_list_field(DataType::Int32, true)
);

pub fn new_dict( name: impl Into<String>, data_type: DataType, nullable: bool, dict_id: i64, dict_is_ordered: bool ) -> Field

Creates a new field that has additional dictionary information

pub fn new_dictionary( name: impl Into<String>, key: DataType, value: DataType, nullable: bool ) -> Field

Create a new Field with DataType::Dictionary

Use Self::new_dict for more advanced dictionary options

Panics

Panics if !key.is_dictionary_key_type

pub fn new_struct( name: impl Into<String>, fields: impl Into<Fields>, nullable: bool ) -> Field

Create a new Field with DataType::Struct

name: the name of the DataType::Struct field

fields: the description of each struct element

nullable: if the DataType::Struct array is nullable

pub fn new_list( name: impl Into<String>, value: impl Into<Arc<Field>>, nullable: bool ) -> Field

Create a new Field with DataType::List

name: the name of the DataType::List field

value: the description of each list element

nullable: if the DataType::List array is nullable

pub fn new_large_list( name: impl Into<String>, value: impl Into<Arc<Field>>, nullable: bool ) -> Field

Create a new Field with DataType::LargeList

name: the name of the DataType::LargeList field

value: the description of each list element

nullable: if the DataType::LargeList array is nullable

pub fn new_map( name: impl Into<String>, entries: impl Into<String>, keys: impl Into<Arc<Field>>, values: impl Into<Arc<Field>>, sorted: bool, nullable: bool ) -> Field

Create a new Field with DataType::Map

name: the name of the DataType::Map field

entries: the name of the inner DataType::Struct field

keys: the map keys

values: the map values

sorted: if the DataType::Map array is sorted

nullable: if the DataType::Map array is nullable

pub fn new_union<S, F, T>( name: S, type_ids: T, fields: F, mode: UnionMode ) -> Field
where S: Into<String>, F: IntoIterator, <F as IntoIterator>::Item: Into<Arc<Field>>, T: IntoIterator<Item = i8>,

Create a new Field with DataType::Union

name: the name of the DataType::Union field

type_ids: the union type ids

fields: the union fields

mode: the union mode

pub fn set_metadata(&mut self, metadata: HashMap<String, String>)

Sets the Field’s optional custom metadata.

pub fn with_metadata(self, metadata: HashMap<String, String>) -> Field

Sets the metadata of this Field to be metadata and returns self

pub const fn metadata(&self) -> &HashMap<String, String>

Returns the immutable reference to the Field’s optional custom metadata.

pub const fn name(&self) -> &String

Returns an immutable reference to the Field’s name.

pub fn with_name(self, name: impl Into<String>) -> Field

Set the name of the Field and returns self.

let field = Field::new("c1", DataType::Int64, false)
   .with_name("c2");

assert_eq!(field.name(), "c2");

pub const fn data_type(&self) -> &DataType

Returns an immutable reference to the Field’s DataType.

pub fn with_data_type(self, data_type: DataType) -> Field

Set DataType of the Field and returns self.

let field = Field::new("c1", DataType::Int64, false)
   .with_data_type(DataType::Utf8);

assert_eq!(field.data_type(), &DataType::Utf8);

pub const fn is_nullable(&self) -> bool

Indicates whether this Field supports null values.

pub fn with_nullable(self, nullable: bool) -> Field

Set nullable of the Field and returns self.

let field = Field::new("c1", DataType::Int64, false)
   .with_nullable(true);

assert_eq!(field.is_nullable(), true);

pub const fn dict_id(&self) -> Option<i64>

Returns the dictionary ID, if this is a dictionary type.

pub const fn dict_is_ordered(&self) -> Option<bool>

Returns whether this Field’s dictionary is ordered, if this is a dictionary type.

pub fn try_merge(&mut self, from: &Field) -> Result<(), ArrowError>

Merge this field into self if it is compatible.

Struct fields are merged recursively.

NOTE: self may be updated to a partial / unexpected state in case of merge failure.

Example:

let mut field = Field::new("c1", DataType::Int64, false);
assert!(field.try_merge(&Field::new("c1", DataType::Int64, true)).is_ok());
assert!(field.is_nullable());

pub fn contains(&self, other: &Field) -> bool

Check to see if self is a superset of other field. Superset is defined as:

if nullability doesn’t match, self needs to be nullable

self.metadata is a superset of other.metadata

all other fields are equal

pub fn size(&self) -> usize

Return size of this instance in bytes.

Includes the size of Self.