Struct arrow::datatypes::Schema

pub struct Schema { /* fields omitted */ }

Describes the meta-data of an ordered sequence of relative types.

Note that this information is only part of the meta-data and not part of the physical memory layout.

Implementations

impl Schema

pub fn empty() -> Self

Creates an empty Schema

pub fn new(fields: Vec<Field>) -> Self

Creates a new Schema from a sequence of Field values

Example

let field_a = Field::new("a", DataType::Int64, false);
let field_b = Field::new("b", DataType::Boolean, false);

let schema = Schema::new(vec![field_a, field_b]);

pub const fn new_with_metadata(
fields: Vec<Field>,
metadata: HashMap<String, String>
) -> Self

Creates a new Schema from a sequence of Field values and adds additional metadata in form of key value pairs.

Example

let field_a = Field::new("a", DataType::Int64, false);
let field_b = Field::new("b", DataType::Boolean, false);

let mut metadata: HashMap<String, String> = HashMap::new();
metadata.insert("row_count".to_string(), "100".to_string());

let schema = Schema::new_with_metadata(vec![field_a, field_b], metadata);

pub fn try_merge(schemas: &[Self]) -> Result<Self>

Merge schema into self if it is compatible. Struct fields will be merged recursively.

Example:

use arrow::datatypes::*;

let merged = Schema::try_merge(&vec![
    Schema::new(vec![
        Field::new("c1", DataType::Int64, false),
        Field::new("c2", DataType::Utf8, false),
    ]),
    Schema::new(vec![
        Field::new("c1", DataType::Int64, true),
        Field::new("c2", DataType::Utf8, false),
        Field::new("c3", DataType::Utf8, false),
    ]),
]).unwrap();

assert_eq!(
    merged,
    Schema::new(vec![
        Field::new("c1", DataType::Int64, true),
        Field::new("c2", DataType::Utf8, false),
        Field::new("c3", DataType::Utf8, false),
    ]),
);

pub const fn fields(&self) -> &Vec<Field>

Returns an immutable reference of the vector of Field instances

pub fn field(&self, i: usize) -> &Field

Returns an immutable reference of a specific Field instance selected using an offset within the internal fields vector

pub fn field_with_name(&self, name: &str) -> Result<&Field>

Returns an immutable reference of a specific Field instance selected by name

pub fn fields_with_dict_id(&self, dict_id: i64) -> Vec<&Field>

Returns a vector of immutable references to all Field instances selected by the dictionary ID they use

pub fn index_of(&self, name: &str) -> Result<usize>

Find the index of the column with the given name

pub const fn metadata(&self) -> &HashMap<String, String>

Returns an immutable reference to the Map of custom metadata key-value pairs.

pub fn column_with_name(&self, name: &str) -> Option<(usize, &Field)>

Look up a column by name and return a immutable reference to the column along with it's index

pub fn to_json(&self) -> Value

Generate a JSON representation of the Schema

pub fn from(json: &Value) -> Result<Self>

Parse a Schema definition from a JSON representation