Trait arrow::datatypes::ArrowNumericType

pub trait ArrowNumericType: ArrowPrimitiveType where
    Self::Simd: Add<Output = Self::Simd> + Sub<Output = Self::Simd> + Mul<Output = Self::Simd> + Div<Output = Self::Simd> + Copy,
    Self::SimdMask: BitAnd<Output = Self::SimdMask> + BitOr<Output = Self::SimdMask> + BitAndAssign + BitOrAssign + Not<Output = Self::SimdMask> + Copy, {
    type Simd;
    type SimdMask;
    fn lanes() -> usize;

    fn init(value: Self::Native) -> Self::Simd;

    fn load(slice: &[Self::Native]) -> Self::Simd;

    fn mask_init(value: bool) -> Self::SimdMask;

    fn mask_from_u64(mask: u64) -> Self::SimdMask;

    fn mask_to_u64(mask: &Self::SimdMask) -> u64;

    fn mask_get(mask: &Self::SimdMask, idx: usize) -> bool;

    fn mask_set(mask: Self::SimdMask, idx: usize, value: bool) -> Self::SimdMask;

    fn mask_select(
        mask: Self::SimdMask, 
        a: Self::Simd, 
        b: Self::Simd
    ) -> Self::Simd;

    fn mask_any(mask: Self::SimdMask) -> bool;

    fn bin_op<F: Fn(Self::Simd, Self::Simd) -> Self::Simd>(
        left: Self::Simd, 
        right: Self::Simd, 
        op: F
    ) -> Self::Simd;

    fn eq(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn ne(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn lt(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn le(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn gt(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn ge(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn write(simd_result: Self::Simd, slice: &mut [Self::Native]);

    fn unary_op<F: Fn(Self::Simd) -> Self::Simd>(
        a: Self::Simd, 
        op: F
    ) -> Self::Simd;
}

A subtype of primitive type that represents numeric values.

SIMD operations are defined in this trait if available on the target system.

Associated Types

type Simd

Defines the SIMD type that should be used for this numeric type

type SimdMask

Defines the SIMD Mask type that should be used for this numeric type

Loading content...
Required methods

fn lanes() -> usize

The number of SIMD lanes available

fn init(value: Self::Native) -> Self::Simd

Initializes a SIMD register to a constant value

fn load(slice: &[Self::Native]) -> Self::Simd

Loads a slice into a SIMD register

fn mask_init(value: bool) -> Self::SimdMask

Creates a new SIMD mask for this SIMD type filling it with value

fn mask_from_u64(mask: u64) -> Self::SimdMask

Creates a new SIMD mask for this SIMD type from the lower-most bits of the given mask. The number of bits used corresponds to the number of lanes of this type

fn mask_to_u64(mask: &Self::SimdMask) -> u64

Creates a bitmask from the given SIMD mask. Each bit corresponds to one vector lane, starting with the least-significant bit.

fn mask_get(mask: &Self::SimdMask, idx: usize) -> bool

Gets the value of a single lane in a SIMD mask

fn mask_set(mask: Self::SimdMask, idx: usize, value: bool) -> Self::SimdMask

Sets the value of a single lane of a SIMD mask

fn mask_select(mask: Self::SimdMask, a: Self::Simd, b: Self::Simd) -> Self::Simd

Selects elements of a and b using mask

fn mask_any(mask: Self::SimdMask) -> bool

Returns true if any of the lanes in the mask are true

fn bin_op<F: Fn(Self::Simd, Self::Simd) -> Self::Simd>(
left: Self::Simd,
right: Self::Simd,
op: F
) -> Self::Simd

Performs a SIMD binary operation

fn eq(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of equal

fn ne(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of not equal

fn lt(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of less than

fn le(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of less than or equal to

fn gt(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of greater than

fn ge(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of greater than or equal to

fn write(simd_result: Self::Simd, slice: &mut [Self::Native])

Writes a SIMD result back to a slice

fn unary_op<F: Fn(Self::Simd) -> Self::Simd>(a: Self::Simd, op: F) -> Self::Simd

Loading content...