arrow::error

Enum ArrowError

pub enum ArrowError {

    NotYetImplemented(String),
    ExternalError(Box<dyn Error + Sync + Send>),
    CastError(String),
    MemoryError(String),
    ParseError(String),
    SchemaError(String),
    ComputeError(String),
    DivideByZero,
    ArithmeticOverflow(String),
    CsvError(String),
    JsonError(String),
    IoError(String, Error),
    IpcError(String),
    InvalidArgumentError(String),
    ParquetError(String),
    CDataInterface(String),
    DictionaryKeyOverflowError,
    RunEndIndexOverflowError,

}

Many different operations in the arrow crate return this error type.

Variants

NotYetImplemented(String)

Returned when functionality is not yet available.

ExternalError(Box<dyn Error + Sync + Send>)

Wraps an external error.

CastError(String)

Error during casting from one type to another.

MemoryError(String)

Memory or buffer error.

ParseError(String)

Error during parsing from a string.

SchemaError(String)

Error during schema-related operations.

ComputeError(String)

Error during computation.

DivideByZero

Error during division by zero.

ArithmeticOverflow(String)

Error when an arithmetic operation overflows.

CsvError(String)

Error during CSV-related operations.

JsonError(String)

Error during JSON-related operations.

IoError(String, Error)

Error during IO operations.

IpcError(String)

Error during IPC operations in arrow-ipc or arrow-flight.

InvalidArgumentError(String)

Error indicating that an unexpected or bad argument was passed to a function.

ParquetError(String)

Error during Parquet operations.

CDataInterface(String)

Error during import or export to/from the C Data Interface

DictionaryKeyOverflowError

Error when a dictionary key is bigger than the key type

RunEndIndexOverflowError

Error when the run end index in a REE array is bigger than the array length

Implementations

impl ArrowError

pub fn from_external_error(error: Box<dyn Error + Sync + Send>) -> ArrowError

Wraps an external error in an ArrowError.