Struct arrow::ffi::ArrowArray

pub struct ArrowArray {
    pub(crate) array: Arc<FFI_ArrowArray>,
    pub(crate) schema: Arc<FFI_ArrowSchema>,
}

Struct used to move an Array from and to the C Data Interface. Its main responsibility is to expose functionality that requires both FFI_ArrowArray and FFI_ArrowSchema.

Import from the C Data Interface

ArrowArray::new to create an array from FFI_ArrowArray and FFI_ArrowSchema

Export to the C Data Interface

Use FFI_ArrowArray and FFI_ArrowSchema directly

Safety

This struct assumes that the incoming data agrees with the C data interface.

Fields
array: Arc<FFI_ArrowArray>schema: Arc<FFI_ArrowSchema>
Implementations

impl ArrowArray

pub fn new(array: FFI_ArrowArray, schema: FFI_ArrowSchema) -> Self

Creates a new ArrowArray from the provided array and schema

pub fn try_new(data: ArrayData) -> Result<Self>

creates a new ArrowArray. This is used to export to the C Data Interface.

Memory Leaks

This method releases buffers. Consumers of this struct must call release before releasing this struct, or contents in buffers leak.

pub unsafe fn empty() -> Self

creates a new empty ArrowArray. Used to import from the C Data Interface.

Safety

See safety of ArrowArray