Function arrow::ipc::gen::File::size_prefixed_root_as_footer_with_opts

pub fn size_prefixed_root_as_footer_with_opts<'b, 'o>(
    opts: &'o VerifierOptions, 
    buf: &'b [u8]
) -> Result<Footer<'b>, InvalidFlatbuffer>

Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed Footer and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_footer_unchecked.