Struct arrow::ipc::gen::SparseTensor::SparseMatrixIndexCSX

pub struct SparseMatrixIndexCSX<'a> {
    pub _tab: Table<'a>,
}

Compressed Sparse format, that is matrix-specific.

Fields
_tab: Table<'a>
Implementations

impl<'a> SparseMatrixIndexCSX<'a>

pub fn init_from_table(table: Table<'a>) -> Self

pub fn create<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr>(
_fbb: &'mut_bldr mut FlatBufferBuilder<'bldr>,
args: &'args SparseMatrixIndexCSXArgs<'args>
) -> WIPOffset<SparseMatrixIndexCSX<'bldr>>

pub const VT_COMPRESSEDAXIS: VOffsetT = 4u16

pub const VT_INDPTRTYPE: VOffsetT = 6u16

pub const VT_INDPTRBUFFER: VOffsetT = 8u16

pub const VT_INDICESTYPE: VOffsetT = 10u16

pub const VT_INDICESBUFFER: VOffsetT = 12u16

pub fn compressedAxis(&self) -> SparseMatrixCompressedAxis

Which axis, row or column, is compressed

pub fn indptrType(&self) -> Int<'a>

The type of values in indptrBuffer

pub fn indptrBuffer(&self) -> &'a Buffer

indptrBuffer stores the location and size of indptr array that represents the range of the rows. The i-th row spans from indptr[i] to indptr[i+1] in the data. The length of this array is 1 + (the number of rows), and the type of index value is long.

For example, let X be the following 6x4 matrix:

  X := [[0, 1, 2, 0],
        [0, 0, 3, 0],
        [0, 4, 0, 5],
        [0, 0, 0, 0],
        [6, 0, 7, 8],
        [0, 9, 0, 0]].

The array of non-zero values in X is:

  values(X) = [1, 2, 3, 4, 5, 6, 7, 8, 9].

And the indptr of X is:

  indptr(X) = [0, 2, 3, 5, 5, 8, 10].

pub fn indicesType(&self) -> Int<'a>

The type of values in indicesBuffer

pub fn indicesBuffer(&self) -> &'a Buffer

indicesBuffer stores the location and size of the array that contains the column indices of the corresponding non-zero values. The type of index value is long.

For example, the indices of the above X is:

  indices(X) = [1, 2, 2, 1, 3, 0, 2, 3, 1].

Note that the indices are sorted in lexicographical order for each row.