Function arrow::ipc::gen::Tensor::root_as_tensor

pub fn root_as_tensor(buf: &[u8]) -> Result<Tensor<'_>, InvalidFlatbuffer>

Verifies that a buffer of bytes contains a Tensor and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_tensor_unchecked.