Type Alias arrow::tensor::Decimal256Tensor

pub type Decimal256Tensor<'a> = Tensor<'a, Decimal256Type>;

Aliased Type

struct Decimal256Tensor<'a> {
    data_type: DataType,
    buffer: Buffer,
    shape: Option<Vec<usize, Global>>,
    strides: Option<Vec<usize, Global>>,
    names: Option<Vec<&'a str, Global>>,
    _marker: PhantomData<Decimal256Type>,
}

Fields
data_type: DataTypebuffer: Buffershape: Option<Vec<usize, Global>>strides: Option<Vec<usize, Global>>names: Option<Vec<&'a str, Global>>_marker: PhantomData<Decimal256Type>
Implementations

impl<'a, T: ArrowPrimitiveType> Tensor<'a, T>

pub fn try_new( buffer: Buffer, shape: Option<Vec<usize>>, strides: Option<Vec<usize>>, names: Option<Vec<&'a str>> ) -> Result<Self>

Creates a new Tensor

pub fn new_row_major( buffer: Buffer, shape: Option<Vec<usize>>, names: Option<Vec<&'a str>> ) -> Result<Self>

Creates a new Tensor using row major memory layout

pub fn new_column_major( buffer: Buffer, shape: Option<Vec<usize>>, names: Option<Vec<&'a str>> ) -> Result<Self>

Creates a new Tensor using column major memory layout

pub fn data_type(&self) -> &DataType

The data type of the Tensor

pub fn shape(&self) -> Option<&Vec<usize>>

The sizes of the dimensions

pub fn data(&self) -> &Buffer

Returns a reference to the underlying Buffer

pub fn strides(&self) -> Option<&Vec<usize>>

The number of bytes between elements in each dimension

pub fn names(&self) -> Option<&Vec<&'a str>>

The names of the dimensions

pub fn ndim(&self) -> usize

The number of dimensions

pub fn dim_name(&self, i: usize) -> Option<&'a str>

The name of dimension i

pub fn size(&self) -> usize

The total number of elements in the Tensor

pub fn is_contiguous(&self) -> Result<bool>

Indicates if the data is laid out contiguously in memory

pub fn is_row_major(&self) -> Result<bool>

Indicates if the memory layout row major

pub fn is_column_major(&self) -> Result<bool>

Indicates if the memory layout column major