arrow::util::bench_util

Function create_longer_string_array_with_same_prefix

pub fn create_longer_string_array_with_same_prefix<Offset: OffsetSizeTrait>(
    size: usize,
    null_density: f32,
) -> GenericStringArray<Offset>

Creates longer string array with same prefix, the prefix should be larger than 4 bytes, and the string length should be larger than 12 bytes so that we can compare the performance with StringViewArray, because StringViewArray has 4 bytes inline for view