arrow::util::bit_chunk_iterator

Struct UnalignedBitChunk

pub struct UnalignedBitChunk<'a> {
    lead_padding: usize,
    trailing_padding: usize,
    prefix: Option<u64>,
    chunks: &'a [u64],
    suffix: Option<u64>,
}

Iterates over an arbitrarily aligned byte buffer

Yields an iterator of aligned u64, along with the leading and trailing u64 necessary to align the buffer to a 8-byte boundary

This is unlike BitChunkIterator which only exposes a trailing u64, and consequently has to perform more work for each read

Fields
lead_padding: usizetrailing_padding: usizeprefix: Option<u64>chunks: &'a [u64]suffix: Option<u64>
Implementations

impl<'a> UnalignedBitChunk<'a>

pub fn new(buffer: &'a [u8], offset: usize, len: usize) -> UnalignedBitChunk<'a>

Create a from a byte array, and and an offset and length in bits

pub fn lead_padding(&self) -> usize

Returns the number of leading padding bits

pub fn trailing_padding(&self) -> usize

Returns the number of trailing padding bits

pub fn prefix(&self) -> Option<u64>

Returns the prefix, if any

pub fn suffix(&self) -> Option<u64>

Returns the suffix, if any

pub fn chunks(&self) -> &'a [u64]

Returns reference to the chunks

pub fn iter( &self, ) -> Chain<Chain<IntoIter<u64>, Cloned<Iter<'a, u64>>>, IntoIter<u64>>

Returns an iterator over the chunks

pub fn count_ones(&self) -> usize

Counts the number of ones