arrow::util::bit_mask

Function set_bits

pub fn set_bits(
    write_data: &mut [u8],
    data: &[u8],
    offset_write: usize,
    offset_read: usize,
    len: usize,
) -> usize

Util function to set bits in a slice of bytes.

This will sets all bits on write_data in the range [offset_write..offset_write+len] to be equal to the bits in data in the range [offset_read..offset_read+len] returns the number of 0 bits data[offset_read..offset_read+len] offset_write, offset_read, and len are in terms of bits