arrow::util::data_gen

Function create_random_array

pub fn create_random_array(
    field: &Field,
    size: usize,
    null_density: f32,
    true_density: f32,
) -> Result<ArrayRef>

Create a random ArrayRef from a DataType with a length, null density and true density (for BooleanArray).

Arguments

field - The field containing the data type for which to create a random array

size - The number of elements in the generated array

null_density - The approximate fraction of null values in the resulting array (0.0 to 1.0)

true_density - The approximate fraction of true values in boolean arrays (0.0 to 1.0)