arrow::util

Module pretty

Utilities for pretty printing RecordBatches and Arrays.

Note this module is not available unless feature = "prettyprint" is enabled.

Functions

pretty_format_batches

Create a visual representation of RecordBatches

pretty_format_batches_with_options

Create a visual representation of RecordBatches with formatting options.

pretty_format_batches_with_schema

Create a visual representation of RecordBatches with a provided schema.

pretty_format_columns

Create a visual representation of ArrayRef

pretty_format_columns_with_options

Create a visual representation of ArrayRef with formatting options.

print_batches

Prints a visual representation of record batches to stdout

print_columns

Prints a visual representation of a list of column to stdout