Module arrow::util::test_util

Utils to make testing easier

Structs

BadIterator

An iterator that is untruthful about its actual length

Functions

arrow_test_data

Returns the arrow test data directory, which is by default stored in a git submodule rooted at arrow/testing/data.

get_data_dir 🔒

Returns a directory path for finding test data.

get_temp_file

Returns file handle for a temp file in ‘target’ directory with a provided content

parquet_test_data

Returns the parquest test data directory, which is by default stored in a git submodule rooted at arrow/parquest-testing/data.

random_bytes

Returns a vector of size n, filled with randomly generated bytes.

seedable_rng

Returns fixed seedable RNG