Function arrow::util::test_util::parquet_test_data

pub fn parquet_test_data() -> String

Returns the parquest test data directory, which is by default stored in a git submodule rooted at arrow/parquest-testing/data.

The default can be overridden by the optional environment variable PARQUET_TEST_DATA

panics when the directory can not be found.

Example:

let testdata = arrow::util::test_util::parquet_test_data();
let filename = format!("{}/binary.parquet", testdata);
assert!(std::path::PathBuf::from(filename).exists());