Struct arrow::array::GenericBinaryArray

pub struct GenericBinaryArray<OffsetSize: BinaryOffsetSizeTrait> { /* fields omitted */ }

Implementations

impl<OffsetSize: BinaryOffsetSizeTrait> GenericBinaryArray<OffsetSize>

pub fn value_offset(&self, i: usize) -> OffsetSize

Returns the offset for the element at index i.

Note this doesn't do any bound checking, for performance reason.

pub fn value_length(&self, i: usize) -> OffsetSize

Returns the length for the element at index i.

Note this doesn't do any bound checking, for performance reason.

pub fn value_offsets(&self) -> Buffer

Returns a clone of the value offset buffer

pub fn value_data(&self) -> Buffer

Returns a clone of the value data buffer

pub fn value(&self, i: usize) -> &[u8]

Returns the element at index i as a byte slice.

pub fn from_vec(v: Vec<&[u8]>) -> Self

Creates a GenericBinaryArray from a vector of byte slices

pub fn from_opt_vec(v: Vec<Option<&[u8]>>) -> Self

Creates a GenericBinaryArray from a vector of Optional (null) byte slices

impl<'a, T: BinaryOffsetSizeTrait> GenericBinaryArray<T>

pub fn iter(&'a self) -> GenericBinaryIter<'a, T>

constructs a new iterator