Struct arrow::array::PrimitiveArray

pub struct PrimitiveArray<T: ArrowPrimitiveType> { /* fields omitted */ }

Array whose elements are of primitive types.

Implementations

impl<T: ArrowPrimitiveType> PrimitiveArray<T>

pub fn len(&self) -> usize

Returns the length of this array.

pub fn is_empty(&self) -> bool

Returns whether this array is empty.

pub fn values(&self) -> &[T::Native]

Returns a slice of the values of this array

pub fn builder(capacity: usize) -> PrimitiveBuilder<T>

pub unsafe fn value_unchecked(&self, i: usize) -> T::Native

Returns the primitive value at index i.

Safety

caller must ensure that the passed in offset is less than the array len()

pub fn value(&self, i: usize) -> T::Native

Returns the primitive value at index i.

Panics of offset i is out of bounds

pub fn from_iter_values<I: IntoIterator<Item = T::Native>>(iter: I) -> Self

Creates a PrimitiveArray based on an iterator of values without nulls

pub fn from_value(value: T::Native, count: usize) -> Self

Creates a PrimitiveArray based on a constant value with count elements

impl<T: ArrowTemporalType + ArrowNumericType> PrimitiveArray<T> where
i64: From<T::Native>,

pub fn value_as_datetime(&self, i: usize) -> Option<NaiveDateTime>

Returns value as a chrono NaiveDateTime, handling time resolution

If a data type cannot be converted to NaiveDateTime, a None is returned. A valid value is expected, thus the user should first check for validity.

pub fn value_as_date(&self, i: usize) -> Option<NaiveDate>

Returns value as a chrono NaiveDate by using Self::datetime()

If a data type cannot be converted to NaiveDate, a None is returned

pub fn value_as_time(&self, i: usize) -> Option<NaiveTime>

Returns a value as a chrono NaiveTime

Date32 and Date64 return UTC midnight as they do not have time resolution

pub fn value_as_duration(&self, i: usize) -> Option<Duration>

Returns a value as a chrono Duration

If a data type cannot be converted to Duration, a None is returned

impl<'a, T: ArrowPrimitiveType> PrimitiveArray<T>

pub fn iter(&'a self) -> PrimitiveIter<'a, T>

constructs a new iterator

impl<T: ArrowPrimitiveType> PrimitiveArray<T>

pub unsafe fn from_trusted_len_iter<I, P>(iter: I) -> Self where
P: Borrow<Option<<T as ArrowPrimitiveType>::Native>>,
I: IntoIterator<Item = P>,

Creates a PrimitiveArray from an iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

impl<T: ArrowTimestampType> PrimitiveArray<T>

pub fn from_vec(data: Vec<i64>, timezone: Option<String>) -> Self

Construct a timestamp array from a vec of i64 values and an optional timezone

impl<T: ArrowTimestampType> PrimitiveArray<T>

pub fn from_opt_vec(data: Vec<Option<i64>>, timezone: Option<String>) -> Self

Construct a timestamp array from a vec of Option values and an optional timezone