Trait arrow::array::ArrayBuilder

pub trait ArrayBuilder: Any {
    pub fn len(&self) -> usize;

    pub fn is_empty(&self) -> bool;

    pub fn finish(&mut self) -> ArrayRef;

    pub fn as_any(&self) -> &dyn Any;

    pub fn as_any_mut(&mut self) -> &mut dyn Any;

    pub fn into_box_any(self: Box<Self>) -> Box<dyn Any>;
}

Trait for dealing with different array builders at runtime

Required methods

pub fn len(&self) -> usize

Returns the number of array slots in the builder

pub fn is_empty(&self) -> bool

Returns whether number of array slots is zero

pub fn finish(&mut self) -> ArrayRef

Builds the array

pub fn as_any(&self) -> &dyn Any

Returns the builder as a non-mutable Any reference.

This is most useful when one wants to call non-mutable APIs on a specific builder type. In this case, one can first cast this into a Any, and then use downcast_ref to get a reference on the specific builder.

pub fn as_any_mut(&mut self) -> &mut dyn Any

Returns the builder as a mutable Any reference.

This is most useful when one wants to call mutable APIs on a specific builder type. In this case, one can first cast this into a Any, and then use downcast_mut to get a reference on the specific builder.

pub fn into_box_any(self: Box<Self>) -> Box<dyn Any>

Returns the boxed builder as a box of Any.

Loading content...