arrow::array

Trait BinaryLikeArrayBuilder

pub trait BinaryLikeArrayBuilder: ArrayBuilder {
    // Required methods
    fn type_name() -> &'static str;
    fn with_capacity(capacity: usize) -> Self;
    fn append_value(&mut self, value: &[u8]);
    fn append_null(&mut self);
}

Trait for binary-like array builders

This trait provides unified interface for builders that append binary-like data such as GenericBinaryBuilder<O> and crate::builder::BinaryViewBuilder

Required Methods

fn type_name() -> &'static str

Returns a human-readable type name for the builder.

fn with_capacity(capacity: usize) -> Self

Creates a new builder with the given row capacity.

fn append_value(&mut self, value: &[u8])

Appends a non-null string value to the builder.

fn append_null(&mut self)

Appends a null value to the builder.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.