Module arrow::array::types

Zero-sized types used to parameterize generic array implementations

Structs

BinaryViewType

BinaryViewType for string arrays

BooleanType

A boolean datatype

Date32Type

A 32-bit date type representing the elapsed time since UNIX epoch in days(32 bits).

Date64Type

A 64-bit date type representing the elapsed time since UNIX epoch in milliseconds(64 bits).

Decimal128Type

The decimal type for a Decimal128Array

Decimal256Type

The decimal type for a Decimal256Array

DurationMicrosecondType

An elapsed time type in microseconds.

DurationMillisecondType

An elapsed time type in milliseconds.

DurationNanosecondType

An elapsed time type in nanoseconds.

DurationSecondType

An elapsed time type in seconds.

Float16Type

A 16-bit floating point number type.

Float32Type

A 32-bit floating point number type.

Float64Type

A 64-bit floating point number type.

GenericBinaryType

ByteArrayType for binary arrays

GenericStringType

ByteArrayType for string arrays

Int8Type

A signed 8-bit integer type.

Int16Type

A signed 16-bit integer type.

Int32Type

A signed 32-bit integer type.

Int64Type

A signed 64-bit integer type.

IntervalDayTime

Value of an IntervalDayTime array

IntervalDayTimeType

A “calendar” interval type representing days and milliseconds. See IntervalDayTime for more details.

IntervalMonthDayNano

Value of an IntervalMonthDayNano array

IntervalMonthDayNanoType

A “calendar” interval type representing months, days, and nanoseconds. See IntervalMonthDayNano for more details.

IntervalYearMonthType

A 32-bit “calendar” interval type representing the number of whole months.

StringViewType

ByteViewType for string arrays

Time32MillisecondType

A 32-bit time type representing the elapsed time since midnight in milliseconds.

Time32SecondType

A 32-bit time type representing the elapsed time since midnight in seconds.

Time64MicrosecondType

A 64-bit time type representing the elapsed time since midnight in microseconds.

Time64NanosecondType

A 64-bit time type representing the elapsed time since midnight in nanoseconds.

TimestampMicrosecondType

A timestamp microsecond type with an optional timezone.

TimestampMillisecondType

A timestamp millisecond type with an optional timezone.

TimestampNanosecondType

A timestamp nanosecond type with an optional timezone.

TimestampSecondType

A timestamp second type with an optional timezone.

UInt8Type

An unsigned 8-bit integer type.

UInt16Type

An unsigned 16-bit integer type.

UInt32Type

An unsigned 32-bit integer type.

UInt64Type

An unsigned 64-bit integer type.

Traits

ArrowDictionaryKeyType

A subtype of primitive type that represents legal dictionary keys. See https://arrow.apache.org/docs/format/Columnar.html

ArrowPrimitiveType

Trait for primitive values, bridging the dynamic-typed nature of Arrow (via DataType) with the static-typed nature of rust types (ArrowNativeType) for all types that implement ArrowNativeType.

ArrowTemporalType

A subtype of primitive type that represents temporal values.

ArrowTimestampType

A timestamp type allows us to create array builders that take a timestamp.

ByteArrayType

A trait over the variable-size byte array types

ByteViewType

A trait over the variable length bytes view array types

DecimalType

A trait over the decimal types, used by PrimitiveArray to provide a generic implementation across the various decimal types

RunEndIndexType

A subtype of primitive type that is used as run-ends index in RunArray. See https://arrow.apache.org/docs/format/Columnar.html

Functions

validate_decimal_precision_and_scale

Validate that precision and scale are valid for T

Type Aliases

BinaryType

An arrow binary array with i32 offsets

LargeBinaryType

An arrow binary array with i64 offsets

LargeUtf8Type

An arrow utf8 array with i64 offsets

Utf8Type

An arrow utf8 array with i32 offsets