Module arrow::buffer

This module contains two main structs: Buffer and MutableBuffer. A buffer represents a contiguous memory region that can be shared via offsets.

Structs

Buffer

Buffer represents a contiguous memory region that can be shared with other buffers and across thread boundaries.

MutableBuffer

A MutableBuffer is Arrow’s interface to build a Buffer out of items or slices of items. Buffers created from MutableBuffer (via into) are guaranteed to have its pointer aligned along cache lines and in multiple of 64 bytes. Use MutableBuffer::push to insert an item, MutableBuffer::extend_from_slice to insert many items, and into to convert it to Buffer.