Struct arrow::buffer::Buffer

pub struct Buffer { /* fields omitted */ }

Buffer is a contiguous memory region of fixed size and is aligned at a 64-byte boundary. Buffer is immutable.

Implementations

impl Buffer

pub unsafe fn from_raw_parts(
ptr: NonNull<u8>,
len: usize,
capacity: usize
) -> Self

Creates a buffer from an existing memory region (must already be byte-aligned), this Buffer will free this piece of memory when dropped.

Arguments

ptr - Pointer to raw parts

len - Length of raw parts in bytes

capacity - Total allocated memory for the pointer ptr, in bytes

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len bytes. If the ptr and capacity come from a Buffer, then this is guaranteed.

pub unsafe fn from_unowned(
ptr: NonNull<u8>,
len: usize,
data: Arc<FFI_ArrowArray>
) -> Self

Creates a buffer from an existing memory region (must already be byte-aligned), this Buffer does not free this piece of memory when dropped.

Arguments

ptr - Pointer to raw parts

len - Length of raw parts in bytes

data - An ffi::FFI_ArrowArray with the data

Safety

This function is unsafe as there is no guarantee that the given pointer is valid for len bytes and that the foreign deallocator frees the region.

pub fn len(&self) -> usize

Returns the number of bytes in the buffer

pub fn capacity(&self) -> usize

Returns the capacity of this buffer. For exernally owned buffers, this returns zero

pub fn is_empty(&self) -> bool

Returns whether the buffer is empty.

pub fn as_slice(&self) -> &[u8]

Returns the byte slice stored in this buffer

pub fn slice(&self, offset: usize) -> Self

Returns a new Buffer that is a slice of this buffer starting at offset. Doing so allows the same memory region to be shared between buffers.

Panics

Panics iff offset is larger than len.

pub fn as_ptr(&self) -> *const u8

Returns a pointer to the start of this buffer.

Note that this should be used cautiously, and the returned pointer should not be stored anywhere, to avoid dangling pointers.

pub unsafe fn typed_data<T: ArrowNativeType + Num>(&self) -> &[T]

View buffer as typed slice.

Safety

ArrowNativeType is public so that it can be used as a trait bound for other public components, such as the ToByteSlice trait. However, this means that it can be implemented by user defined types, which it is not intended for.

Also typed_data::<bool> is unsafe as 0x00 and 0x01 are the only valid values for bool in Rust. However, bool arrays in Arrow are bit-packed which breaks this condition.

pub fn bit_slice(&self, offset: usize, len: usize) -> Self

Returns a slice of this buffer starting at a certain bit offset. If the offset is byte-aligned the returned buffer is a shallow clone, otherwise a new buffer is allocated and filled with a copy of the bits in the range.

pub fn bit_chunks(&self, offset: usize, len: usize) -> BitChunks<'_>

Returns a BitChunks instance which can be used to iterate over this buffers bits in larger chunks and starting at arbitrary bit offsets. Note that both offset and length are measured in bits.

pub fn count_set_bits(&self) -> usize

Returns the number of 1-bits in this buffer.

pub fn count_set_bits_offset(&self, offset: usize, len: usize) -> usize

Returns the number of 1-bits in this buffer, starting from offset with length bits inspected. Note that both offset and length are measured in bits.