arrow::compute

Module kernels

Computation kernels on Arrow Arrays

Modules

aggregate

Defines aggregations over Arrow arrays.

arity

Kernels for operating on PrimitiveArrays

bitwise

Module contains bitwise operations on arrays

boolean

Defines boolean kernels on Arrow BooleanArray’s, e.g. AND, OR and NOT.

cast

Cast kernels to convert ArrayRef between supported datatypes.

cast_utils

Parser implementations for converting strings to Arrow types

cmp

Comparison kernels for Arrays.

comparison

Comparison kernels for Arrays.

concat

Defines concat kernel for ArrayRef

concat_elements

Provides utility functions for concatenation of elements in arrays.

filter

Defines filter kernels

interleave

Interleave elements from multiple arrays

length

Defines kernel for length of string arrays and binary arrays

nullif

Implements the nullif function for Arrow arrays.

numeric

Defines numeric arithmetic kernels on PrimitiveArray, such as add

partition

Defines partition kernel for ArrayRef

rank

Provides rank function to assign a rank to each value in an array

regexp

Defines kernel to extract substrings based on a regular expression of a [Large]StringArray

sort

Defines sort kernel for ArrayRef

substring

Defines kernel to extract a substring of an Array Supported array types: GenericStringArray, GenericBinaryArray, FixedSizeBinaryArray, DictionaryArray

take

Defines take kernel for Array

temporal

Defines temporal kernels for time and date related functions.

union_extract

Defines union_extract kernel for UnionArray

window

Defines windowing functions, like shifting

zip

zip: Combine values from two arrays based on boolean mask

Functions

cast

Cast array to the provided data type and return a new Array with type to_type, if possible.