Module arrow::compute::kernels

Computation kernels on Arrow Arrays

Modules

aggregate

Defines aggregations over Arrow arrays.

arithmetic

Defines basic arithmetic kernels for PrimitiveArrays.

arity

Defines kernels suitable to perform operations to primitive arrays.

boolean

Defines boolean kernels on Arrow BooleanArray’s, e.g. AND, OR and NOT.

cast

Defines cast kernels for ArrayRef, to convert Arrays between supported datatypes.

cast_utils

comparison

Defines basic comparison kernels for PrimitiveArrays.

concat

Defines concat kernel for ArrayRef

filter

Defines miscellaneous array kernels.

length

Defines kernel for length of a string array

limit

Defines miscellaneous array kernels.

partition

Defines partition kernel for ArrayRef

regexp

Defines kernel to extract substrings based on a regular expression of a [Large]StringArray

sort

Defines sort kernel for ArrayRef

substring

Defines kernel to extract a substring of a [Large]StringArray

take

Defines take kernel for Array

temporal

Defines temporal kernels for time and date related functions.

window

Defines windowing functions, like shifting

zip