Module arrow::compute::kernels::arithmetic

Defines basic arithmetic kernels for PrimitiveArrays.

These kernels can leverage SIMD if available on your system. Currently no runtime detection is provided, you should enable the specific SIMD intrinsics using RUSTFLAGS="-C target-feature=+avx2" for example. See the documentation here for more information.

Functions

add
Perform left + right operation on two arrays. If either left or right value is null then the result is also null.

divide
Perform left / right operation on two arrays. If either left or right value is null then the result is also null. If any right hand value is zero then the result of this operation will be Err(ArrowError::DivideByZero).

math_op
Helper function to perform math lambda function on values from two arrays. If either left or right value is null then the output value is also null, so 1 + null is null.

multiply
Perform left * right operation on two arrays. If either left or right value is null then the result is also null.

negate
Perform - operation on an array. If value is null then the result is also null.

signed_unary_math_op
Helper function to perform math lambda function on values from single array of signed numeric type. If value is null then the output value is also null, so -null is null.

subtract
Perform left - right operation on two arrays. If either left or right value is null then the result is also null.