Module arrow::compute::kernels::arity

Defines kernels suitable to perform operations to primitive arrays.

Functions

binary

Given two arrays of length len, calls op(a[i], b[i]) for i in 0..len, collecting the results in a PrimitiveArray. If any index is null in either a or b, the corresponding index in the result will also be null

binary_mut

Given two arrays of length len, calls op(a[i], b[i]) for i in 0..len, mutating the mutable PrimitiveArray a. If any index is null in either a or b, the corresponding index in the result will also be null.

try_binary

Applies the provided fallible binary operation across a and b, returning any error, and collecting the results into a PrimitiveArray. If any index is null in either a or b, the corresponding index in the result will also be null

try_binary_mut

Applies the provided fallible binary operation across a and b by mutating the mutable PrimitiveArray a with the results, returning any error. If any index is null in either a or b, the corresponding index in the result will also be null

try_unary

See PrimitiveArray::try_unary

try_unary_dynDeprecated

Applies a fallible unary function to an array with primitive values.

try_unary_mut

See PrimitiveArray::try_unary_mut

unary

See PrimitiveArray::unary

unary_dynDeprecated

Applies an infallible unary function to an array with primitive values.

unary_mut

See PrimitiveArray::unary_mut