Module arrow::compute::kernels::arity

Defines kernels suitable to perform operations to primitive arrays.

Functions

unary
Applies an unary and infalible function to a primitive array. This is the fastest way to perform an operation on a primitive array when the benefits of a vectorized operation outweights the cost of branching nulls and non-nulls.