Module arrow::compute::kernels::comparison

Comparison kernels for Arrays.

Functions

contains

Perform SQL CONTAINS(left, right)

ends_with

Perform SQL ENDSWITH(left, right)

eqDeprecated

Perform left == right operation on two PrimitiveArrays.

eq_binaryDeprecated

Perform left == right operation on BinaryArray / LargeBinaryArray.

eq_binary_scalarDeprecated

Perform left == right operation on BinaryArray / LargeBinaryArray and a scalar

eq_boolDeprecated

Perform left == right operation on BooleanArray

eq_bool_scalarDeprecated

Perform left == right operation on BooleanArray and a scalar

eq_dynDeprecated

Perform left == right operation on two (dynamic) Arrays.

eq_dyn_binary_scalarDeprecated

Perform left == right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

eq_dyn_bool_scalarDeprecated

Perform left == right operation on an array and a numeric scalar value.

eq_dyn_scalarDeprecated

Perform left == right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

eq_dyn_utf8_scalarDeprecated

Perform left == right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

eq_scalarDeprecated

Perform left == right operation on a PrimitiveArray and a scalar value.

eq_utf8Deprecated

Perform left == right operation on StringArray / LargeStringArray.

eq_utf8_scalarDeprecated

Perform left == right operation on StringArray / LargeStringArray and a scalar.

gtDeprecated

Perform left > right operation on two PrimitiveArrays. Non-null values are greater than null values.

gt_binaryDeprecated

Perform left > right operation on BinaryArray / LargeBinaryArray.

gt_binary_scalarDeprecated

Perform left > right operation on BinaryArray / LargeBinaryArray and a scalar.

gt_boolDeprecated

Perform left > right operation on BooleanArray

gt_bool_scalarDeprecated

Perform left > right operation on BooleanArray and a scalar

gt_dynDeprecated

Perform left > right operation on two (dynamic) Arrays.

gt_dyn_binary_scalarDeprecated

Perform left > right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

gt_dyn_bool_scalarDeprecated

Perform left > right operation on an array and a numeric scalar value. Supports BooleanArrays.

gt_dyn_scalarDeprecated

Perform left > right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

gt_dyn_utf8_scalarDeprecated

Perform left > right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

gt_eqDeprecated

Perform left >= right operation on two PrimitiveArrays. Non-null values are greater than null values.

gt_eq_binaryDeprecated

Perform left >= right operation on BinaryArray / LargeBinaryArray.

gt_eq_binary_scalarDeprecated

Perform left >= right operation on BinaryArray / LargeBinaryArray and a scalar.

gt_eq_boolDeprecated

Perform left >= right operation on BooleanArray

gt_eq_bool_scalarDeprecated

Perform left >= right operation on BooleanArray and a scalar

gt_eq_dynDeprecated

Perform left >= right operation on two (dynamic) Arrays.

gt_eq_dyn_binary_scalarDeprecated

Perform left >= right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

gt_eq_dyn_bool_scalarDeprecated

Perform left >= right operation on an array and a numeric scalar value. Supports BooleanArrays.

gt_eq_dyn_scalarDeprecated

Perform left >= right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

gt_eq_dyn_utf8_scalarDeprecated

Perform left >= right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

gt_eq_scalarDeprecated

Perform left >= right operation on a PrimitiveArray and a scalar value. Non-null values are greater than null values.

gt_eq_utf8Deprecated

Perform left >= right operation on StringArray / LargeStringArray.

gt_eq_utf8_scalarDeprecated

Perform left >= right operation on StringArray / LargeStringArray and a scalar.

gt_scalarDeprecated

Perform left > right operation on a PrimitiveArray and a scalar value. Non-null values are greater than null values.

gt_utf8Deprecated

Perform left > right operation on StringArray / LargeStringArray.

gt_utf8_scalarDeprecated

Perform left > right operation on StringArray / LargeStringArray and a scalar.

ilike

Perform SQL left ILIKE right

in_list

Checks if a GenericListArray contains a value in the PrimitiveArray

in_list_utf8

Checks if a GenericListArray contains a value in the GenericStringArray

like

Perform SQL left LIKE right

ltDeprecated

Perform left < right operation on two PrimitiveArrays. Null values are less than non-null values.

lt_binaryDeprecated

Perform left < right operation on BinaryArray / LargeBinaryArray.

lt_binary_scalarDeprecated

Perform left < right operation on BinaryArray / LargeBinaryArray and a scalar.

lt_boolDeprecated

Perform left < right operation on BooleanArray

lt_bool_scalarDeprecated

Perform left < right operation on BooleanArray and a scalar

lt_dynDeprecated

Perform left < right operation on two (dynamic) Arrays.

lt_dyn_binary_scalarDeprecated

Perform left < right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

lt_dyn_bool_scalarDeprecated

Perform left < right operation on an array and a numeric scalar value. Supports BooleanArrays.

lt_dyn_scalarDeprecated

Perform left < right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

lt_dyn_utf8_scalarDeprecated

Perform left < right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

lt_eqDeprecated

Perform left <= right operation on two PrimitiveArrays. Null values are less than non-null values.

lt_eq_binaryDeprecated

Perform left <= right operation on BinaryArray / LargeBinaryArray.

lt_eq_binary_scalarDeprecated

Perform left <= right operation on BinaryArray / LargeBinaryArray and a scalar.

lt_eq_boolDeprecated

Perform left <= right operation on BooleanArray

lt_eq_bool_scalarDeprecated

Perform left <= right operation on BooleanArray and a scalar

lt_eq_dynDeprecated

Perform left <= right operation on two (dynamic) Arrays.

lt_eq_dyn_binary_scalarDeprecated

Perform left <= right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

lt_eq_dyn_bool_scalarDeprecated

Perform left <= right operation on an array and a numeric scalar value. Supports BooleanArrays.

lt_eq_dyn_scalarDeprecated

Perform left <= right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

lt_eq_dyn_utf8_scalarDeprecated

Perform left <= right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

lt_eq_scalarDeprecated

Perform left <= right operation on a PrimitiveArray and a scalar value. Null values are less than non-null values.

lt_eq_utf8Deprecated

Perform left <= right operation on StringArray / LargeStringArray.

lt_eq_utf8_scalarDeprecated

Perform left <= right operation on StringArray / LargeStringArray and a scalar.

lt_scalarDeprecated

Perform left < right operation on a PrimitiveArray and a scalar value. Null values are less than non-null values.

lt_utf8Deprecated

Perform left < right operation on StringArray / LargeStringArray.

lt_utf8_scalarDeprecated

Perform left < right operation on StringArray / LargeStringArray and a scalar.

neqDeprecated

Perform left != right operation on two PrimitiveArrays.

neq_binaryDeprecated

Perform left != right operation on BinaryArray / LargeBinaryArray.

neq_binary_scalarDeprecated

Perform left != right operation on BinaryArray / LargeBinaryArray and a scalar.

neq_boolDeprecated

Perform left != right operation on BooleanArray

neq_bool_scalarDeprecated

Perform left != right operation on BooleanArray and a scalar

neq_dynDeprecated

Perform left != right operation on two (dynamic) Arrays.

neq_dyn_binary_scalarDeprecated

Perform left != right operation on an array and a numeric scalar value. Supports BinaryArray and LargeBinaryArray

neq_dyn_bool_scalarDeprecated

Perform left != right operation on an array and a numeric scalar value. Supports BooleanArrays.

neq_dyn_scalarDeprecated

Perform left != right operation on an array and a numeric scalar value. Supports PrimitiveArrays, and DictionaryArrays that have primitive values.

neq_dyn_utf8_scalarDeprecated

Perform left != right operation on an array and a numeric scalar value. Supports StringArrays, and DictionaryArrays that have string values

neq_scalarDeprecated

Perform left != right operation on a PrimitiveArray and a scalar value.

neq_utf8Deprecated

Perform left != right operation on StringArray / LargeStringArray.

neq_utf8_scalarDeprecated

Perform left != right operation on StringArray / LargeStringArray and a scalar.

nilike

Perform SQL left NOT ILIKE right

nlike

Perform SQL left NOT LIKE right

regexp_is_match_utf8

Perform SQL array ~ regex_array operation on StringArray / LargeStringArray. If regex_array element has an empty value, the corresponding result value is always true.

regexp_is_match_utf8_scalar

Perform SQL array ~ regex_array operation on StringArray / LargeStringArray and a scalar.

starts_with

Perform SQL STARTSWITH(left, right)

unary_cmpDeprecated

Applies an unary and infallible comparison function to a primitive array.