Module arrow::compute::kernels::comparison

Defines basic comparison kernels for PrimitiveArrays.

These kernels can leverage SIMD if available on your system. Currently no runtime detection is provided, you should enable the specific SIMD intrinsics using RUSTFLAGS="-C target-feature=+avx2" for example. See the documentation here for more information.

Functions

contains

Checks if a GenericListArray contains a value in the PrimitiveArray

contains_utf8

Checks if a GenericListArray contains a value in the GenericStringArray

eq

Perform left == right operation on two arrays.

eq_scalar

Perform left == right operation on an array and a scalar value.

eq_utf8

Perform left == right operation on StringArray / LargeStringArray.

eq_utf8_scalar

Perform left == right operation on StringArray / LargeStringArray and a scalar.

gt

Perform left > right operation on two arrays. Non-null values are greater than null values.

gt_eq

Perform left >= right operation on two arrays. Non-null values are greater than null values.

gt_eq_scalar

Perform left >= right operation on an array and a scalar value. Non-null values are greater than null values.

gt_eq_utf8

Perform left >= right operation on StringArray / LargeStringArray.

gt_eq_utf8_scalar

Perform left >= right operation on StringArray / LargeStringArray and a scalar.

gt_scalar

Perform left > right operation on an array and a scalar value. Non-null values are greater than null values.

gt_utf8

Perform left > right operation on StringArray / LargeStringArray.

gt_utf8_scalar

Perform left > right operation on StringArray / LargeStringArray and a scalar.

like_utf8

Perform SQL left LIKE right operation on StringArray / LargeStringArray.

like_utf8_scalar

Perform SQL left LIKE right operation on StringArray / LargeStringArray and a scalar.

lt

Perform left < right operation on two arrays. Null values are less than non-null values.

lt_eq

Perform left <= right operation on two arrays. Null values are less than non-null values.

lt_eq_scalar

Perform left <= right operation on an array and a scalar value. Null values are less than non-null values.

lt_eq_utf8

Perform left <= right operation on StringArray / LargeStringArray.

lt_eq_utf8_scalar

Perform left <= right operation on StringArray / LargeStringArray and a scalar.

lt_scalar

Perform left < right operation on an array and a scalar value. Null values are less than non-null values.

lt_utf8

Perform left < right operation on StringArray / LargeStringArray.

lt_utf8_scalar

Perform left < right operation on StringArray / LargeStringArray and a scalar.

neq

Perform left != right operation on two arrays.

neq_scalar

Perform left != right operation on an array and a scalar value.

neq_utf8

Perform left != right operation on StringArray / LargeStringArray.

neq_utf8_scalar

Perform left != right operation on StringArray / LargeStringArray and a scalar.

nlike_utf8

Perform SQL left NOT LIKE right operation on StringArray / LargeStringArray.

nlike_utf8_scalar

Perform SQL left NOT LIKE right operation on StringArray / LargeStringArray and a scalar.

no_simd_compare_op

Evaluate op(left, right) for PrimitiveArrays using a specified comparison function.

no_simd_compare_op_scalar

Evaluate op(left, right) for PrimitiveArray and scalar using a specified comparison function.