Trait arrow::datatypes::ArrowNativeType

pub trait ArrowNativeType: Debug + Send + Sync + Copy + PartialOrd + FromStr + Default + JsonSerializable {
    fn from_usize(_: usize) -> Option<Self> { ... }

    fn to_usize(&self) -> Option<usize> { ... }

    fn from_i32(_: i32) -> Option<Self> { ... }

    fn from_i64(_: i64) -> Option<Self> { ... }
}

Trait expressing a Rust type that has the same in-memory representation as Arrow. This includes i16, f32, but excludes bool (which in arrow is represented in bits). In little endian machines, types that implement ArrowNativeType can be memcopied to arrow buffers as is.

Provided methods

fn from_usize(_: usize) -> Option<Self>

Convert native type from usize.

fn to_usize(&self) -> Option<usize>

Convert native type to usize.

fn from_i32(_: i32) -> Option<Self>

Convert native type from i32.

fn from_i64(_: i64) -> Option<Self>

Convert native type from i64.

Loading content...