Trait arrow::datatypes::ArrowPrimitiveType

pub trait ArrowPrimitiveType: 'static {
    type Native: ArrowNativeType;

    const DATA_TYPE: DataType;
    fn get_byte_width() -> usize { ... }

    fn default_value() -> Self::Native { ... }
}

Trait indicating a primitive fixed-width type (bool, ints and floats).

Associated Types

type Native: ArrowNativeType

Corresponding Rust native type for the primitive type.

Loading content...
Associated Constants

const DATA_TYPE: DataType

the corresponding Arrow data type of this primitive type.

Loading content...
Provided methods

fn get_byte_width() -> usize

Returns the byte width of this primitive type.

fn default_value() -> Self::Native

Returns a default value of this primitive type.

This is useful for aggregate array ops like sum(), mean().

Loading content...