Trait arrow::datatypes::ArrowSignedNumericType

pub trait ArrowSignedNumericType: ArrowNumericType where
    Self::SignedSimd: Neg<Output = Self::SignedSimd>, {
    type SignedSimd;
    fn load_signed(slice: &[Self::Native]) -> Self::SignedSimd;

    fn signed_unary_op<F: Fn(Self::SignedSimd) -> Self::SignedSimd>(
        a: Self::SignedSimd, 
        op: F
    ) -> Self::SignedSimd;

    fn write_signed(simd_result: Self::SignedSimd, slice: &mut [Self::Native]);
}

A subtype of primitive type that represents signed numeric values.

SIMD operations are defined in this trait if available on the target system.

Associated Types

type SignedSimd

Defines the SIMD type that should be used for this numeric type

Required methods

fn load_signed(slice: &[Self::Native]) -> Self::SignedSimd

Loads a slice of signed numeric type into a SIMD register

fn signed_unary_op<F: Fn(Self::SignedSimd) -> Self::SignedSimd>(
a: Self::SignedSimd,
op: F
) -> Self::SignedSimd

Performs a SIMD unary operation on signed numeric type

fn write_signed(simd_result: Self::SignedSimd, slice: &mut [Self::Native])

Writes a signed SIMD result back to a slice