Trait arrow::ffi::ArrowArrayRef

pub trait ArrowArrayRef {
    // Required methods
    fn owner(&self) -> &Arc<FFI_ArrowArray>;
    fn array(&self) -> &FFI_ArrowArray;
    fn schema(&self) -> &FFI_ArrowSchema;
    fn data_type(&self) -> Result<DataType>;

    // Provided methods
    fn to_data(&self) -> Result<ArrayData> { ... }
    fn buffers(&self, can_contain_null_mask: bool) -> Result<Vec<Buffer>> { ... }
    fn buffer_len(&self, i: usize) -> Result<usize> { ... }
    fn null_bit_buffer(&self) -> Option<Buffer> { ... }
    fn child(&self, index: usize) -> ArrowArrayChild<'_> { ... }
    fn dictionary(&self) -> Option<ArrowArrayChild<'_>> { ... }
}

Required Methods

fn owner(&self) -> &Arc<FFI_ArrowArray>

fn array(&self) -> &FFI_ArrowArray

fn schema(&self) -> &FFI_ArrowSchema

fn data_type(&self) -> Result<DataType>

Provided Methods

fn to_data(&self) -> Result<ArrayData>

fn buffers(&self, can_contain_null_mask: bool) -> Result<Vec<Buffer>>

returns all buffers, as organized by Rust (i.e. null buffer is skipped if it’s present in the spec of the type)

fn buffer_len(&self, i: usize) -> Result<usize>

Returns the length, in bytes, of the buffer i (indexed according to the C data interface) Rust implementation uses fixed-sized buffers, which require knowledge of their len. for variable-sized buffers, such as the second buffer of a stringArray, we need to fetch offset buffer’s len to build the second buffer.

fn null_bit_buffer(&self) -> Option<Buffer>

returns the null bit buffer. Rust implementation uses a buffer that is not part of the array of buffers. The C Data interface’s null buffer is part of the array of buffers.

fn child(&self, index: usize) -> ArrowArrayChild<'_>

fn dictionary(&self) -> Option<ArrowArrayChild<'_>>