Trait arrow::ffi::ArrowArrayRef

pub trait ArrowArrayRef {
    fn owner(&self) -> &Arc<FFI_ArrowArray>;

    fn array(&self) -> &FFI_ArrowArray;

    fn schema(&self) -> &FFI_ArrowSchema;

    fn data_type(&self) -> Result<DataType>;

    fn to_data(&self) -> Result<ArrayData> { ... }

    fn buffers(&self) -> Result<Vec<Buffer>> { ... }

    fn buffer_len(&self, i: usize) -> Result<usize> { ... }

    fn null_bit_buffer(&self) -> Option<Buffer> { ... }

    fn child(&self, index: usize) -> ArrowArrayChild<'_> { ... }
}

Required methods

fn owner(&self) -> &Arc<FFI_ArrowArray>

fn array(&self) -> &FFI_ArrowArray

fn schema(&self) -> &FFI_ArrowSchema

fn data_type(&self) -> Result<DataType>

Provided methods

fn to_data(&self) -> Result<ArrayData>

fn buffers(&self) -> Result<Vec<Buffer>>

returns all buffers, as organized by Rust (i.e. null buffer is skipped)

fn buffer_len(&self, i: usize) -> Result<usize>

Returns the length, in bytes, of the buffer i (indexed according to the C data interface)

fn null_bit_buffer(&self) -> Option<Buffer>

returns the null bit buffer. Rust implementation uses a buffer that is not part of the array of buffers. The C Data interface’s null buffer is part of the array of buffers.

fn child(&self, index: usize) -> ArrowArrayChild<'_>