Module arrow::ffi_stream

Contains declarations to bind to the C Stream Interface.

This module has two main interfaces: One interface maps C ABI to native Rust types, i.e. convert c-pointers, c_char, to native rust. This is handled by FFI_ArrowArrayStream.

The second interface is used to import FFI_ArrowArrayStream as Rust implementation RecordBatch reader. This is handled by ArrowArrayStreamReader.

// create an record batch reader natively
let file = File::open("arrow_file").unwrap();
let reader = Box::new(FileReader::try_new(file).unwrap());

// export it
let mut stream = FFI_ArrowArrayStream::empty();
unsafe { export_reader_into_raw(reader, &mut stream) };

// consumed and used by something else...

// import it
let stream_reader = unsafe { ArrowArrayStreamReader::from_raw(&mut stream).unwrap() };
let imported_schema = stream_reader.schema();

let mut produced_batches = vec![];
for batch in stream_reader {
     produced_batches.push(batch.unwrap());
}
Ok(())
}

Structs

ArrowArrayStreamReader

A RecordBatchReader which imports Arrays from FFI_ArrowArrayStream. Struct used to fetch RecordBatch from the C Stream Interface. Its main responsibility is to expose RecordBatchReader functionality that requires FFI_ArrowArrayStream.

ExportedArrayStream 🔒

FFI_ArrowArrayStream

ABI-compatible struct for ArrayStream from C Stream Interface See https://arrow.apache.org/docs/format/CStreamInterface.html#structure-definitions This was created by bindgen

StreamPrivateData 🔒

Constants

EINVAL 🔒

EIO 🔒

ENOMEM 🔒

ENOSYS 🔒

Functions

export_reader_into_raw⚠Deprecated

Exports a record batch reader to raw pointer of the C Stream Interface provided by the consumer.

get_error_code 🔒

get_last_error 🔒 ⚠

get_next 🔒 ⚠

get_schema 🔒 ⚠

get_stream_schema 🔒

Gets schema from a raw pointer of FFI_ArrowArrayStream. This is used when constructing ArrowArrayStreamReader to cache schema.

release_stream 🔒 ⚠