Module arrow::ipc::writer

Arrow IPC File and Stream Writers

The FileWriter and StreamWriter have similar interfaces, however the FileWriter expects a reader that supports Seeking

Structs

DictionaryTracker
Keeps track of dictionaries that have been written, to avoid emitting the same dictionary multiple times. Can optionally error if an update to an existing dictionary is attempted, which isn't allowed in the FileWriter.

EncodedData
Stores the encoded data, which is an ipc::Message, and optional Arrow data

FileWriter

IpcDataGenerator

IpcWriteOptions
IPC write options used to control the behaviour of the writer

StreamWriter

Functions

write_message
Write a message's IPC data and buffers, returning metadata and buffer data lengths written