Trait arrow::record_batch::RecordBatchReader

pub trait RecordBatchReader: Iterator<Item = Result<RecordBatch>> {
    fn schema(&self) -> SchemaRef;

    fn next_batch(&mut self) -> Result<Option<RecordBatch>> { ... }
}

Trait for types that can read RecordBatch's.

Required methods

fn schema(&self) -> SchemaRef

Returns the schema of this RecordBatchReader.

Implementation of this trait should guarantee that all RecordBatch's returned by this reader should have the same schema as returned from this method.

Loading content...
Provided methods

fn next_batch(&mut self) -> Result<Option<RecordBatch>>

👎 Deprecated since 2.0.0:
This method is deprecated in favour of next from the trait Iterator.

Reads the next RecordBatch.

Loading content...