arrow
Module tensor

Arrow Tensor Type, defined in format/Tensor.fbs.

Structs

Tensor

Tensor of primitive types

Functions

compute_column_major_strides 🔒

Computes the strides required assuming a column major memory layout

compute_row_major_strides 🔒

Computes the strides required assuming a row major memory layout

Type Aliases

BooleanTensor

Tensor of type BooleanType

Date32Tensor

Tensor of type Int8Type

Date64Tensor

Tensor of type Int16Type

Decimal128Tensor

Tensor of type Decimal128Type

Decimal256Tensor

Tensor of type Decimal256Type

DurationMicrosecondTensor

Tensor of type DurationMicrosecondType

DurationMillisecondTensor

Tensor of type DurationMillisecondType

DurationNanosecondTensor

Tensor of type DurationNanosecondType

DurationSecondTensor

Tensor of type DurationSecondType

Float16Tensor

Tensor of type Float16Type

Float32Tensor

Tensor of type Float32Type

Float64Tensor

Tensor of type Float64Type

Int8Tensor

Tensor of type Int8Type

Int16Tensor

Tensor of type Int16Type

Int32Tensor

Tensor of type Int32Type

Int64Tensor

Tensor of type Int64Type

IntervalDayTimeTensor

Tensor of type IntervalDayTimeType

IntervalMonthDayNanoTensor

Tensor of type IntervalMonthDayNanoType

IntervalYearMonthTensor

Tensor of type IntervalYearMonthType

Time32MillisecondTensor

Tensor of type Time32MillisecondType

Time32SecondTensor

Tensor of type Time32SecondType

Time64MicrosecondTensor

Tensor of type Time64MicrosecondType

Time64NanosecondTensor

Tensor of type Time64NanosecondType

TimestampMicrosecondTensor

Tensor of type TimestampMicrosecondType

TimestampMillisecondTensor

Tensor of type TimestampMillisecondType

TimestampNanosecondTensor

Tensor of type TimestampNanosecondType

TimestampSecondTensor

Tensor of type TimestampSecondType

UInt8Tensor

Tensor of type UInt8Type

UInt16Tensor

Tensor of type UInt16Type

UInt32Tensor

Tensor of type UInt32Type

UInt64Tensor

Tensor of type UInt64Type