Module arrow::tensor

Arrow Tensor Type, defined in format/Tensor.fbs.

Structs

Tensor

Tensor of primitive types

Functions

compute_column_major_strides 🔒

Computes the strides required assuming a column major memory layout

compute_row_major_strides 🔒

Computes the strides required assuming a row major memory layout

Type Definitions

BooleanTensor

Date32Tensor

Date64Tensor

Decimal128Tensor

Decimal256Tensor

DurationMicrosecondTensor

DurationMillisecondTensor

DurationNanosecondTensor

DurationSecondTensor

Float16Tensor

Float32Tensor

Float64Tensor

Int8Tensor

Int16Tensor

Int32Tensor

Int64Tensor

IntervalDayTimeTensor

IntervalMonthDayNanoTensor

IntervalYearMonthTensor

Time32MillisecondTensor

Time32SecondTensor

Time64MicrosecondTensor

Time64NanosecondTensor

TimestampMicrosecondTensor

TimestampMillisecondTensor

TimestampNanosecondTensor

TimestampSecondTensor

UInt8Tensor

UInt16Tensor

UInt32Tensor

UInt64Tensor