Module arrow::util

Modules

bench_util

Utils to make benchmarking easier

bit_chunk_iterator

bit_util

Utils for working with bits

data_gen

Utilities to generate random arrays and batches

display

Functions for printing array values, as strings, for debugging purposes. See the pretty crate for additional functions for record batch pretty printing.

integration_util

Utils for JSON integration testing

string_writer

String Writer This string writer encapsulates std::string::String and implements std::io::Write trait, which makes String as a writable object like File.

test_util

Utils to make testing easier