Module arrow::util::bench_util

Utils to make benchmarking easier

Functions

create_binary_array

Creates an random (but fixed-seeded) binary array of a given size and null density

create_boolean_array

Creates an random (but fixed-seeded) array of a given size and null density

create_dict_from_values

Creates a random (but fixed-seeded) dictionary array of a given size and null density with the provided values array

create_fsb_array

Creates an random (but fixed-seeded) array of a given size and null density

create_primitive_array

Creates an random (but fixed-seeded) array of a given size and null density

create_primitive_array_with_seed

create_primitive_run_array

Create primitive run array for given logical and physical array lengths

create_sparse_dict_from_values

Creates a random (but fixed-seeded) dictionary array of a given size and null density with the provided values array and key range

create_string_array

Creates an random (but fixed-seeded) array of a given size and null density

create_string_array_for_runs

Create string array to be used by run array builder. The string array will result in run array with physical length of physical_array_len and logical length of logical_array_len

create_string_array_with_len

Creates a random (but fixed-seeded) array of a given size, null density and length

create_string_dict_array

Creates an random (but fixed-seeded) array of a given size and null density consisting of random 4 character alphanumeric strings