Module arrow::util::bit_util

Utils for working with bits

Functions

ceil
Returns the ceil of value/divisor

get_bit
Returns whether bit at position i in data is set or not

get_bit_raw⚠
Returns whether bit at position i in data is set or not.

round_upto_multiple_of_64
Returns the nearest number that is >= than num and is a multiple of 64

round_upto_power_of_2
Returns the nearest multiple of factor that is >= than num. Here factor must be a power of 2.

set_bit
Sets bit at position i for data

set_bit_raw⚠
Sets bit at position i for data

unset_bit
Sets bit at position i for data to 0

unset_bit_raw⚠
Sets bit at position i for data to 0