arrow::util

Module data_gen

Utilities to generate random arrays and batches

Traits

RandomTemporalValue

Useful for testing. The range of values are not likely to be representative of the actual bounds.

Functions

create_random_array

Create a random ArrayRef from a DataType with a length, null density and true density (for BooleanArray).

create_random_batch

Create a random RecordBatch from a schema

create_random_list_array ğŸ”’

create_random_map_array ğŸ”’

create_random_null_buffer ğŸ”’

create_random_offsets ğŸ”’

Generate random offsets for list arrays

create_random_struct_array ğŸ”’

create_random_temporal_array ğŸ”’