arrow::util
Module display

Functions for printing array values as human-readable strings.

This is often used for debugging or logging purposes.

See the pretty crate for additional functions for record batch pretty printing.

Structs

ArrayFormatter

A string formatter for an Array

FormatOptions

Options for formatting arrays

ValueFormatter

Implements Display for a specific array value

Enums

DurationFormat

Format for displaying durations

Functions

array_value_to_string

Get the value at the given row in an array as a String.

lexical_to_string

Converts numeric type to a String