arrow::util::display

Struct FormatOptions

pub struct FormatOptions<'a> {
    safe: bool,
    null: &'a str,
    date_format: Option<&'a str>,
    datetime_format: Option<&'a str>,
    timestamp_format: Option<&'a str>,
    timestamp_tz_format: Option<&'a str>,
    time_format: Option<&'a str>,
    duration_format: DurationFormat,
}

Options for formatting arrays

By default nulls are formatted as "" and temporal types formatted according to RFC3339

Fields
safe: boolnull: &'a strdate_format: Option<&'a str>datetime_format: Option<&'a str>timestamp_format: Option<&'a str>timestamp_tz_format: Option<&'a str>time_format: Option<&'a str>duration_format: DurationFormat
Implementations

impl<'a> FormatOptions<'a>

pub const fn new() -> FormatOptions<'a>

Creates a new set of format options

pub const fn with_display_error(self, safe: bool) -> FormatOptions<'a>

If set to true any formatting errors will be written to the output instead of being converted into a std::fmt::Error

pub const fn with_null(self, null: &'a str) -> FormatOptions<'a>

Overrides the string used to represent a null

Defaults to ""

pub const fn with_date_format( self, date_format: Option<&'a str>, ) -> FormatOptions<'a>

Overrides the format used for DataType::Date32 columns

pub const fn with_datetime_format( self, datetime_format: Option<&'a str>, ) -> FormatOptions<'a>

Overrides the format used for DataType::Date64 columns

pub const fn with_timestamp_format( self, timestamp_format: Option<&'a str>, ) -> FormatOptions<'a>

Overrides the format used for DataType::Timestamp columns without a timezone

pub const fn with_timestamp_tz_format( self, timestamp_tz_format: Option<&'a str>, ) -> FormatOptions<'a>

Overrides the format used for DataType::Timestamp columns with a timezone

pub const fn with_time_format( self, time_format: Option<&'a str>, ) -> FormatOptions<'a>

Overrides the format used for DataType::Time32 and DataType::Time64 columns

pub const fn with_duration_format( self, duration_format: DurationFormat, ) -> FormatOptions<'a>

Overrides the format used for duration columns

Defaults to DurationFormat::ISO8601