Module arrow::util::pretty

Utilities for pretty printing record batches. Note this module is not available unless feature = "prettyprint" is enabled.

Functions

pretty_format_batches

Create a visual representation of record batches

pretty_format_batches_with_options

Create a visual representation of record batches

pretty_format_columns

Create a visual representation of columns

pretty_format_columns_with_options

print_batches

Prints a visual representation of record batches to stdout

print_columns

Prints a visual representation of a list of column to stdout